### 2020-03-01 #2128 @labarba
@openjournals/dev Can you check why our bot didn't run `cloc` on starting this pre-review? And also why it's flagging all those missing DOIs when the paper only has two citations.
