### 2020-03-01 #2128 @labarba
👋 @kyleniemeyer — The authors suggested you as handling editor for this submission. The software facilitates acquisition, analysis and fitting of pendant drop profiles to obtain surface and interfacial tension. Can you take it?
