### 2020-03-03 #2128 @arfon
> @openjournals/dev Can you check why our bot didn't run cloc on starting this pre-review?

Not sure why that didn't work the first time.

> And also why it's flagging all those missing DOIs when the paper only has two citations.

The DOI checker just parses whatever is in the `paper.bib` file. This BibTeX file looks to have a whole bunch of entries that aren't being cited in the paper?
