### 2020-03-06 #2128 @kyleniemeyer
Hi @jdber1, thanks for your interest in JOSS.

My first question is on how easy/hard it will be for others to review this software—will reviewers be limited to those with access to certain experimental equipment? If so, then I'll really need your help to find folks who can review.
