### 2020-03-10 #2128 @jdber1
Hi @kyleniemeyer, it would probably be better for the reviewers to have an experimental setup, but it is not necessary. I can give you a list of people who I know have used OpenDrop v1 for their research. Do they need to be active on github to review? That might be a bit harder.

There are also test images for interfacial tension in tests/samples/images, and I can add test images for contact angle too, for those without experimental apparatus.
