### 2020-03-25 #2128 @kyleniemeyer
Hi @jdber1, sorry for the delay and thanks for the info.

I think adding some test images, and setting up associated tests, is a great idea. I also welcome suggestions for reviewers! They don't have to be active on GitHub to be asked to review, though completing a review does require a GitHub account. I'm happy to reach out to folks via email. Feel free to give names here, or email me if you prefer.
