### 2020-03-25 #2128 @jdber1
No problems at all @kyleniemeyer. Potential reviewers include:

Stuart Prescott -  [llimeht](https://github.com/llimeht)
Trent Parker - [tmpchem](https://github.com/tmpchem)
Marcus Hanwell- [cryos](https://github.com/cryos)
Palli Thordarson - contact details [here](http://www.chemistry.unsw.edu.au/research/research-groups/thordarson-group)
Jan Vermant - contact details [here](https://mavt.ethz.ch/the-department/people/person-detail.MjAzNTI5.TGlzdC8xMzc1LC02OTM2MTkzMA==.html)

I will also get some test images so that the reviewers can use the software without experimental apparatus.
