### 2020-04-13 #2128 @kyleniemeyer
Hi @llimeht @tmpchem and @cryos, would any of you be interested in reviewing this submission for JOSS?

I'm not sure if you've all reviewed for JOSS before (or recently), but we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The package we would ask you to review is ["OpenDrop: Open-source software for pendant drop tensiometry & contact angle measurements"](https://github.com/jdber1/opendrop) and there is a [short associated paper](https://github.com/openjournals/joss-papers/blob/joss.02128/joss.02128/10.21105.joss.02128.pdf) to read too.

Thanks!
