### 2020-04-16 430f454
Allow more image file types to be selectable in image acquisition file browser. Also fix incorrect jpeg MIME type.
