### 2020-04-16 d2deb95
Redesign main menu
