+++ setup.py
@@ -31,7 +31,7 @@ from setuptools import find_packages, setup
 
 setup(
     name='opendrop',
-    version='3.1.6dev0',
+    version='3.1.7dev0',
     packages=find_packages(exclude=['tests', 'manual_tests', 'docs']),
     package_data={
         'opendrop.res': ['images/*']
