### 2020-06-15 f19b4c7
Refactor geometry.py (type hints also improved)
