+++ opendrop/app/conan/results/individual/detail/info/component.py
@@ -102,7 +102,7 @@ class InfoView(View['InfoPresenter', Gtk.Widget]):
         self._background_ro.props.pixbuf = pixbuf_from_array(image)
 
         self._render.props.canvas_size = image.shape[1::-1]
-        self._render.viewport_extents = Rect2(pos=(0, 0), size=image.shape[1::-1])
+        self._render.viewport_extents = Rect2(position=(0, 0), size=image.shape[1::-1])
 
     def _do_destroy(self) -> None:
         self._render.destroy()
