+++ opendrop/app/common/image_processing/plugins/define_region/component.py
@@ -160,8 +160,8 @@ class DefineRegionPluginPresenter(Presenter['DefineRegionPluginView']):
             return
 
         self.view.bn_dragging.set(Rect2(
-            p0=self._model.begin_define_pos,
-            p1=current_cursor_pos,
+            pt0=self._model.begin_define_pos,
+            pt1=current_cursor_pos,
         ))
 
     def _do_destroy(self) -> None:
