### 2020-06-27 629fa80
Add new simpler app framework
