+++ opendrop/appfw/_template.py
@@ -1,0 +1,170 @@
+from typing import Type, TypeVar, Callable, Any, Optional, Generic, overload
+import sys
+from gi.repository import Gtk, GObject, Gio
+
+T = TypeVar('T')
+
+WidgetClassT = TypeVar('WidgetClassT', bound=Type[Gtk.Widget])
+WidgetT = TypeVar('WidgetT', bound=Gtk.Widget)
+
+
+def templated(path: str) -> Callable[[WidgetClassT], WidgetClassT]:
+    def wrapper(widget_class: WidgetClassT) -> WidgetClassT:
+        mark_as_templated(widget_class, template_path=path)
+        return widget_class
+
+    return wrapper
+
+
+def mark_as_templated(widget_class: Type[Gtk.Widget], template_path: str) -> None:
+    # If resource name is not an absolute path, we use some tricks to resolve it.
+    if not template_path.startswith('/'):
+        package_name = get_owner_package(widget_class)
+        if package_name is None:
+            raise RuntimeError(
+                "Could not determine the owner package of '{}' to resolve relative template path"
+                .format(widget_class.__name__)
+            )
+
+        template_path = resolve_relative_resource(package_name, template_path)
+
+    template = Gio.resources_lookup_data(template_path, Gio.ResourceLookupFlags.NONE).get_data().decode()
+
+    widget_class.__widget_template__ = TemplateSpec(template)
+
+    # This is a hack that mirrors Gtk.Template's implementation, see:
+    #     https://gitlab.gnome.org/GNOME/pygobject/-/blob/c745866df9e4394c6615e79779fff1221a84953a/gi/_gtktemplate.py#L94
+    # It is essentially a post __init__() hook baked into GObjects.
+    widget_class.__dontuse_ginstance_init__ = template_init
+
+    return widget_class
+
+
+class TemplateSpec:
+    def __init__(self, template: str) -> None:
+        self.template = template
+
+
+class TemplatePrivate:
+    private_key = '_template_private_data'
+
+    def __init__(self, builder: Gtk.Builder) -> None:
+        self.builder = builder
+        self.child_cache = {}
+
+    def install(self, target: Any) -> None:
+        target.__dict__[__class__.private_key] = self
+
+    @staticmethod
+    def get(owner: Any) -> 'TemplatePrivate':
+        return owner.__dict__[__class__.private_key]
+
+
+def template_init(self):
+    template_spec = self.__widget_template__
+
+    # Build template
+    template = template_spec.template
+
+    builder = Gtk.Builder()
+    builder.extend_with_template(self, type(self), template, len(template))
+
+    private = TemplatePrivate(builder)
+    private.install(self)
+
+    # Connect callbacks
+    builder.connect_signals_full(builder_connect_helper, self)
+
+    after_template_init = getattr(self, 'after_template_init', None)
+    if after_template_init:
+        after_template_init()
+
+
+def builder_connect_helper(
+        builder: Gtk.Builder,
+        obj: GObject.Object,
+        signal_name: str,
+        handler_name: str,
+        connect_obj: Optional[GObject.Object],
+        flags: GObject.ConnectFlags,
+        component: Gtk.Widget,
+) -> None:
+    if flags & GObject.ConnectFlags.SWAPPED:
+        raise ValueError(
+            "{!r} not supported"
+            .format(GObject.ConnectFlags.SWAPPED)
+        )
+
+    handler = getattr(component, handler_name)
+    args = ()
+
+    if connect_obj:
+        # Special case handling of connect_obj with g_signal_connect_object() is not needed (and not
+        # implemented anyway) in python.
+        args = (connect_obj,)
+
+    if flags & GObject.ConnectFlags.AFTER:
+        obj.connect_after(signal_name, handler, *args)
+    else:
[…]
