### 2020-07-02 38a880c
Replace vendor/glibcoro with vendor/aioglib
