### 2020-07-02 b1bc975
Add 'injector' dependency
