### 2020-07-02 728d1a2
Don't require 'typing_extensions' if python version >= 3.8
