### 2020-07-02 bc275df
Initial refactor to abandon 'mvp' module
EntryPoint class is no longer used, it is replaced by Gtk.Application
