### 2020-07-02 59e63bd
Refactor MainMenu to use new app framework
Rename 'res' module to 'assets'. The data.gresource bundle will need to be
recompiled each time the assets or .ui templates are changed.
