+++ opendrop/app/main_menu.py
@@ -1,169 +1,19 @@
-# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
-# OpenDrop is released under the GNU GPL License. You are free to
-# modify and distribute the code, but always under the same license
-# (i.e. you cannot make commercial derivatives).
-#
-# If you use this software in your research, please cite the following
-# journal articles:
-#
-# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
-# R. F. Tabor, Measurement of surface and interfacial tension using
-# pendant drop tensiometry. Journal of Colloid and Interface Science 454
-# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
-#
-# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
-# and J. D. Berry, OpenDrop: Open-source software for pendant drop
-# tensiometry & contact angle measurements, submitted to the Journal of
-# Open Source Software
-#
-# These citations help us not only to understand who is using and
-# developing OpenDrop, and for what purpose, but also to justify
-# continued development of this code and other open source resources.
-#
-# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
-# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-# PURPOSE.  See the GNU General Public License for more details.  You
-# should have received a copy of the GNU General Public License along
-# with this software.  If not, see <https://www.gnu.org/licenses/>.
+from gi.repository import Gtk
 
-
-import pkg_resources
-from gi.repository import Gtk, GdkPixbuf, Gio, GLib, Gdk
-
-from opendrop.mvp import ComponentSymbol, View, Presenter
+from opendrop.appfw import Inject, component
 from opendrop.app.services.app import OpendropService
 
-main_menu_cs = ComponentSymbol()  # type: ComponentSymbol[Gtk.Widget]
 
+@component(
+    template_path='./main_menu.ui',
+)
+class MainMenu(Gtk.Window):
+    __gtype_name__ = 'MainMenu'
 
-@main_menu_cs.view()
-class MainMenuView(View['MainMenuPresenter', Gtk.Widget]):
-    OPENDROP_LOGO = GdkPixbuf.Pixbuf.new_from_stream(
-        Gio.MemoryInputStream.new_from_bytes(
-            GLib.Bytes(
-                pkg_resources.resource_stream('opendrop.res', 'images/logo_wide.png').read()
-            )
-        )
-    )
+    app_service = Inject(OpendropService)
 
-    IFT_BUTTON_IMAGE = GdkPixbuf.Pixbuf.new_from_stream(
-        Gio.MemoryInputStream.new_from_bytes(
-            GLib.Bytes(
-                pkg_resources.resource_stream('opendrop.res', 'images/ift_btn.png').read()
-            )
-        )
-    )
+    def hdl_ift_btn_clicked(self, button: Gtk.Button) -> None:
+        self.app_service.goto_ift()
 
-    CONAN_BUTTON_IMAGE = GdkPixbuf.Pixbuf.new_from_stream(
-        Gio.MemoryInputStream.new_from_bytes(
-            GLib.Bytes(
-                pkg_resources.resource_stream('opendrop.res', 'images/conan_btn.png').read()
-            )
-        )
-    )
-
-    def _do_init(self) -> Gtk.Widget:
-        self._window = Gtk.Window(
-            title='OpenDrop',
[…]
