### 2020-07-03 2c5d250
Add PEP 517 build system
With the intention of automating GLib resource compilation.
