+++ pyproject.toml
@@ -1,0 +1,4 @@
+[build-system]
+requires = []
+build-backend = "buildapi"
+backend-path = ["build_backend"]
