+++ build_backend/buildapi.py
@@ -1,6 +1,0 @@
-def build_wheel(wheel_directory, config_settings=None, metadata_directory=None):
-    raise NotImplementedError
-
-
-def build_sdist(sdist_directory, config_settings=None):
-    raise NotImplementedError
