+++ site_scons/site_tools/pydist.py
@@ -1,0 +1,407 @@
+import base64
+from collections import defaultdict
+import csv
+import hashlib
+import io
+import itertools
+from pathlib import Path
+from typing import List, Optional, Any, Mapping
+import zipfile
+
+from SCons.Script import Action, Builder, DirScanner, Scanner
+from SCons.Environment import OverrideEnvironment
+import SCons.Node
+import SCons.Node.FS
+import SCons.Node.Python
+import SCons.Util
+
+
+def exists(env):
+    return True
+
+
+def generate(env):
+    env.AddMethod(wheel_package_builder_wrapper, 'WheelPackage')
+    env.AddMethod(wheel_method, 'Wheel')
+
+    env['BUILDERS']['_WheelFileInternal'] = Builder(
+        action=Action(
+            wheel_package_build,
+            varlist=[
+                'AUTHOR',
+                'NAME',
+                'VERSION',
+                'AUTHOR',
+                'AUTHOR_EMAIL',
+                'LICENSE',
+                'ARCHITECTURE',
+                'PACKAGE_METADATA',
+                'SOURCE_URL',
+                'BUILD',
+                '_WHEEL_ROOT_IS_PURELIB',
+                '_WHEEL_TAG',
+            ],
+            cmdstr="Packaging $TARGET",
+        ),
+        target_factory=env.File,
+        source_factory=lambda s: WheelEntry(s, env.Entry(s)),
+        source_scanner=Scanner(
+            function=wheel_entry_scan,
+            skeys=[WheelEntry.SCANNER_KEY],
+        ),
+    )
+
+
+def wheel_package_builder_wrapper(
+        env,
+        target,
+        source=None,
+        *,
+        purelib=(),
+        platlib=(),
+        scripts=(),
+        packages=None,
+        root_is_purelib=True,
+        python_tag='py3',
+        abi_tag='none',
+        platform_tag='any',
+):
+    if source is None:
+        source = target
+        target = '.'
+
+    env = OverrideEnvironment(env, {
+        '_WHEEL_ROOT_IS_PURELIB': root_is_purelib,
+        '_WHEEL_TAG': '{}-{}-{}'.format(python_tag, abi_tag, platform_tag),
+    })
+
+    wheel_filename = format_wheel_package_filename(**env)
+
+    target = env.arg2nodes(target, node_factory=env.Dir)
+    target = target[0].File(wheel_filename)
+    
+    packages = packages or {}
+
+    source = env.arg2nodes(source, node_factory=env.Entry)
+    purelib = env.arg2nodes(purelib, node_factory=env.Entry)
+    platlib = env.arg2nodes(platlib, node_factory=env.Entry)
+    scripts = env.arg2nodes(scripts, node_factory=env.Entry)
+    packages = {name: env.Dir(d) for name, d in packages.items()}
+
+    wheel_entries = []
+
+    for category, nodes in {None: source, 'purelib': purelib, 'platlib': platlib}.items():
+        for node in nodes:
+            if isinstance(node, WheelEntry):
+                if category is not None:
+                    raise ValueError(
+                        "Cannot use WheelEntry sources with convenience parameters 'purelib' and 'platlib'"
+                    )
+                wheel_entry = node
+            else:
+                arcpath = find_package_entry_dest(node, packages)
+                wheel_entry = WheelEntry(arcpath, node, category=category)
+
+            wheel_entries.append(wheel_entry)
+
+    for node in scripts:
+        wheel_entries.append(WheelEntry(
+            arcpath=node.name,
+            source=node,
+            category='scripts',
+        ))
+
+    return env._WheelFileInternal(target, wheel_entries)
+
+
[…]
