### 2020-07-09 1cbef6c
Use setuptools_scm to manage version numbering
It was a bit laborious to manually increment the version number in the
appropriate source file. Having this process automated shaves off a bit
of effort when it comes to making changes to the code.
