+++ requirements.txt
@@ -1,0 +1,10 @@
+matplotlib
+numpy
+scipy
+pycairo
+pygobject
+pytest
+setuptools
+injector
+importlib_resources; python_version<'3.7'
+typing_extensions; python_version<'3.8'
