### 2020-07-09 dd20bec
Remove setuptools requirement
I don't think it's needed anymore.
