### 2020-07-09 ea8e6bd
Move opendrop/app/__main__.py to opendrop/__main__.py
Also pass return of app.run() to sys.exit()
