### 2020-07-10 58785b2
Don't use __dontuse_ginstance_init__ hack
Use another hack instead.., override class's __init__ with our own.
