### 2020-07-10 f431762
Refactor IFTRoot component to use new appfw
