### 2020-07-10 a5d38e1
Rename @templated to @templatewidget and @component to @componentclass
