### 2020-07-10 2ca5c7f
Make FloatEntry work with GtkBuilder
Property types can't be python objects because GtkBuilder doesn't know
how to demarshal those yet.

Some type hints also fixed (some ints should have been floats).
