### 2020-07-10 38e9c07
Fix a bug where our redefined Injector could not be injected
