### 2020-07-10 b155e44
Refactor IFT physical_parameters to use new appfw
