+++ opendrop/app/ift/physical_parameters/component.py
@@ -1,163 +1,0 @@
-# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
-# OpenDrop is released under the GNU GPL License. You are free to
-# modify and distribute the code, but always under the same license
-# (i.e. you cannot make commercial derivatives).
-#
-# If you use this software in your research, please cite the following
-# journal articles:
-#
-# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
-# R. F. Tabor, Measurement of surface and interfacial tension using
-# pendant drop tensiometry. Journal of Colloid and Interface Science 454
-# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
-#
-# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
-# and J. D. Berry, OpenDrop: Open-source software for pendant drop
-# tensiometry & contact angle measurements, submitted to the Journal of
-# Open Source Software
-#
-# These citations help us not only to understand who is using and
-# developing OpenDrop, and for what purpose, but also to justify
-# continued development of this code and other open source resources.
-#
-# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
-# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-# PURPOSE.  See the GNU General Public License for more details.  You
-# should have received a copy of the GNU General Public License along
-# with this software.  If not, see <https://www.gnu.org/licenses/>.
-
-
-from gi.repository import Gtk
-
-from opendrop.app.common.footer.linearnav import linear_navigator_footer_cs
-from opendrop.app.common.wizard import WizardPageControls
-from opendrop.mvp import ComponentSymbol, View, Presenter
-from opendrop.utility.bindable.gextension import GObjectPropertyBindable
-from opendrop.widgets.float_entry import FloatEntry
-from .model import PhysicalParametersModel
-
-physical_parameters_cs = ComponentSymbol()  # type: ComponentSymbol[Gtk.Widget]
-
-
-@physical_parameters_cs.view(options=['footer_area'])
-class PhysicalParametersView(View['PhysicalParametersPresenter', Gtk.Widget]):
-    def _do_init(self, footer_area: Gtk.Grid) -> Gtk.Widget:
-        self._widget = Gtk.Grid(margin=20, row_spacing=10, column_spacing=10)
-
-        # Label widgets
-        inner_density_lbl = Gtk.Label('Inner density (kg/m³):', xalign=0)
-        self._widget.attach(inner_density_lbl, 0, 0, 1, 1)
-
-        outer_density_lbl = Gtk.Label('Outer density (kg/m³):', xalign=0)
-        self._widget.attach(outer_density_lbl, 0, 1, 1, 1)
-
-        needle_width_lbl = Gtk.Label('Needle diameter (mm):', xalign=0)
-        self._widget.attach(needle_width_lbl, 0, 2, 1, 1)
-
-        gravity_lbl = Gtk.Label('Gravity (m/s²):', xalign=0)
-        self._widget.attach(gravity_lbl, 0, 3, 1, 1)
-
-        # Input widgets
-        self._inner_density_inp = FloatEntry(lower=0, width_chars=10)
-        self._widget.attach_next_to(self._inner_density_inp, inner_density_lbl, Gtk.PositionType.RIGHT, 1, 1)
-
-        self._outer_density_inp = FloatEntry(lower=0, width_chars=10)
-        self._widget.attach_next_to(self._outer_density_inp, outer_density_lbl, Gtk.PositionType.RIGHT, 1, 1)
-
-        self._needle_width_inp = FloatEntry(lower=0, width_chars=10)
[…]
