+++ opendrop/app/ift/analysis/__init__.py
@@ -29,5 +29,4 @@
 
 from .analysis import IFTDropAnalysis
 from .features import FeatureExtractor, FeatureExtractorParams
-from .physical_properties import PhysicalPropertiesCalculator, PhysicalPropertiesCalculatorParams
 from .young_laplace_fit import YoungLaplaceFitter
