+++ opendrop/app/ift/analysis/analysis.py
@@ -41,8 +41,8 @@ from opendrop.utility.bindable import AccessorBindable, VariableBindable
 from opendrop.utility.bindable.typing import Bindable
 from opendrop.utility.geometry import Vector2
 from .features import FeatureExtractor
-from .physical_properties import PhysicalPropertiesCalculator
 from .young_laplace_fit import YoungLaplaceFitter
+from opendrop.app.ift.services.quantities import PhysicalPropertiesCalculator
 
 
 # Classes

[…]
