### 2020-07-10 c3582c6
Refactor IFTImageProcessing to use new appfw
