### 2020-07-14 b7c1bd0
More refactor to use new appfw
