+++ opendrop/app/ift/report/overview/master.py
@@ -1,0 +1,262 @@
+# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
+# OpenDrop is released under the GNU GPL License. You are free to
+# modify and distribute the code, but always under the same license
+# (i.e. you cannot make commercial derivatives).
+#
+# If you use this software in your research, please cite the following
+# journal articles:
+#
+# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
+# R. F. Tabor, Measurement of surface and interfacial tension using
+# pendant drop tensiometry. Journal of Colloid and Interface Science 454
+# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
+#
+# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
+# and J. D. Berry, OpenDrop: Open-source software for pendant drop
+# tensiometry & contact angle measurements, submitted to the Journal of
+# Open Source Software
+#
+# These citations help us not only to understand who is using and
+# developing OpenDrop, and for what purpose, but also to justify
+# continued development of this code and other open source resources.
+#
+# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
+# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+# PURPOSE.  See the GNU General Public License for more details.  You
+# should have received a copy of the GNU General Public License along
+# with this software.  If not, see <https://www.gnu.org/licenses/>.
+
+
+from typing import Optional, Sequence
+
+from gi.repository import Gtk, Pango, GObject
+
+from opendrop.app.ift.analysis import IFTDropAnalysis
+from opendrop.appfw import componentclass, TemplateChild
+
+
+@componentclass(
+    template_path='./master.ui',
+)
+class IFTReportOverviewMaster(Gtk.ScrolledWindow):
+    __gtype_name__ = 'IFTReportOverviewMaster'
+
+    _tree_view = TemplateChild('tree_view')
+    _tree_model = TemplateChild('tree_model')
+    _tree_selection = TemplateChild('tree_selection')
+
+    _initial_selection = None
+    _analyses = ()
+
+    _template_ready = False
+
+    class _RowPresenter:
+        TIMESTAMP_COL = 0
+        STATUS_COL = 1
+        LOG_TEXT_COL = 2
+
+        def __init__(self, row_ref: Gtk.TreeRowReference, analysis: IFTDropAnalysis) -> None:
+            self.analysis = analysis
+            self.row_ref = row_ref
+
+            self._event_connections = [
+                analysis.bn_image_timestamp.on_changed.connect(
+                    self._update_timestamp
+                ),
+                analysis.bn_status.on_changed.connect(
+                    self._update_status_text
+                ),
+                analysis.bn_log.on_changed.connect(
+                    self._update_log_text
+                ),
+            ]
+
+            self._update_timestamp()
+            self._update_status_text()
+            self._update_log_text()
+
+        def _update_timestamp(self) -> None:
+            timestamp = self.analysis.bn_image_timestamp.get()
+            text = format(timestamp, '.1f')
+            self._tree_model.set_value(self._tree_iter, column=self.TIMESTAMP_COL, value=text)
+
+        def _update_status_text(self) -> None:
+            status = self.analysis.bn_status.get()
+            text = status.display_name
+            self._tree_model.set_value(self._tree_iter, column=self.STATUS_COL, value=text)
+
+        def _update_log_text(self) -> None:
+            log_history = self.analysis.bn_log.get()
+            log_lines = log_history.splitlines()
+
[…]
