+++ opendrop/appfw/_template.py
@@ -1,47 +1,50 @@
-from typing import Type, TypeVar, Callable, Any, Optional, Generic, overload
-import sys
-from gi.repository import Gtk, GObject, Gio
+from typing import Any, Generic, TypeVar, Type, overload, Literal
+import xml.etree.ElementTree as ET
+from gi.repository import Gtk, GObject
+
 
 T = TypeVar('T')
-
-WidgetClassT = TypeVar('WidgetClassT', bound=Type[Gtk.Widget])
-WidgetT = TypeVar('WidgetT', bound=Gtk.Widget)
+U = TypeVar('U')
 
 
-def templatewidget(path: str) -> Callable[[WidgetClassT], WidgetClassT]:
-    def wrapper(widget_class: WidgetClassT) -> WidgetClassT:
-        mark_as_templated(widget_class, template_path=path)
-        return widget_class
+class TemplateInfo:
+    def __init__(self, data: str) -> None:
+        self.data = data
+        xml = ET.fromstring(data)
 
-    return wrapper
+        name = None
+        parent = None
+        for e in xml.iter('template'):
+            name = e.get('class')
+            parent = e.get('parent')
+            break
 
+        if name is None or parent is None:
+            raise ValueError("Template tag is missing 'class' or 'parent' attribute")
 
-def mark_as_templated(widget_class: WidgetClassT, template_path: str) -> WidgetClassT:
-    # If resource name is not an absolute path, we use some tricks to resolve it.
-    if not template_path.startswith('/'):
-        package_name = get_owner_package(widget_class)
-        if package_name is None:
-            raise RuntimeError(
-                "Could not determine the owner package of '{}' to resolve relative template path"
-                .format(widget_class.__name__)
-            )
-
-        template_path = resolve_relative_resource(package_name, template_path)
-
-    template = Gio.resources_lookup_data(template_path, Gio.ResourceLookupFlags.NONE).get_data().decode()
-
-    widget_class.__widget_template__ = TemplateSpec(template, widget_class.__init__)
-
-    # Hacky.., override with our own __init__ function.
-    widget_class.__init__ = override_init
-
-    return widget_class
+        self.name = name
+        self.parent = GObject.type_from_name(parent).pytype
 
 
-class TemplateSpec:
-    def __init__(self, template: str, original_init: Callable) -> None:
-        self.template = template
-        self.original_init = original_init
+class TemplateChild(Generic[T]):
+    def __init__(self, name: str) -> None:
+        self._name = name
+
+    @overload
+    def __get__(self, instance, owner) -> T: ...
+    @overload
+    def __get__(self: U, instance: Literal[None], owner) -> U: ...
+
+    def __get__(self, instance, owner=None):
+        if owner is None:
+            return self
+
+        template_priv = TemplatePrivate.get(instance)
+
+        if self._name not in template_priv.child_cache:
+            template_priv.child_cache[self._name] = template_priv.builder.get_object(self._name)
+
+        return template_priv.child_cache[self._name]
 
 
 class TemplatePrivate:
@@ -57,116 +60,3 @@ class TemplatePrivate:
     @staticmethod
     def get(owner: Any) -> 'TemplatePrivate':
         return owner.__dict__[__class__.private_key]
-
-
-def override_init(self, *args, **kwargs) -> None:
-    template_spec = self.__widget_template__
-
-    # Invoke original __init__ function
-    template_spec.original_init(self, *args, **kwargs)
-
-    # Build template
-    template = template_spec.template
-
-    builder = Gtk.Builder()
-    builder.extend_with_template(self, type(self), template, len(template))
-
-    private = TemplatePrivate(builder)
-    private.install(self)
-
-    # Connect callbacks
[…]
