+++ opendrop/appfw/_injector.py
@@ -1,2 +1,0 @@
-from injector import *
-from ._injector_override import *
