+++ opendrop/appfw/_component.py
@@ -1,23 +1,251 @@
-from typing import Sequence, Type, TypeVar, Callable
-from gi.repository import Gtk
+import sys
+from typing import Any, Callable, Sequence, Type, TypeVar, Optional
+from types import MethodType
 
-from ._template import mark_as_templated
-from ._injector import mark_as_injection_container
+from gi.repository import GObject, Gio, Gtk
+from gi.types import GObjectMeta
+
+from ._injector_override import Injector
+from ._template import TemplateInfo, TemplatePrivate
+from ._presenter import Presenter
+
 
 WidgetClassT = TypeVar('WidgetClassT', bound=Type[Gtk.Widget])
+PresenterClassT = TypeVar('PresenterClassT', bound=Type[Presenter])
+T = TypeVar('T')
 
 
-def componentclass(
-        template_path: str = '',
+def component(
+        template_path: str,
         modules: Sequence = (),
-        auto_bind: bool = True,
-) -> Callable[[WidgetClassT], WidgetClassT]:
-    def wrapper(widget_class: WidgetClassT) -> WidgetClassT:
-        mark_as_injection_container(widget_class, modules, auto_bind)
-
-        if template_path:
-            mark_as_templated(widget_class, template_path)
-
-        return widget_class
-
+        auto_bind: bool = True
+) -> Callable[[PresenterClassT], PresenterClassT]:
+    def wrapper(presenter_class: PresenterClassT) -> PresenterClassT:
+        register_component(presenter_class, template_path, modules, auto_bind)
+        return presenter_class
     return wrapper
+
+
+def register_component(
+        presenter_class: Type[Presenter],
+        template_path: str,
+        modules: Sequence,
+        auto_bind: bool
+) -> None:
+    # If resource name is not an absolute path, we use some tricks to resolve it.
+    if not template_path.startswith('/'):
+        package_name = get_owner_package(presenter_class)
+        if package_name is None:
+            raise RuntimeError(
+                "Could not determine the owner package of '{}' to resolve relative template path"
+                .format(presenter_class.__name__)
+            )
+
+        template_path = resolve_relative_resource(package_name, template_path)
+
+    template = Gio.resources_lookup_data(template_path, Gio.ResourceLookupFlags.NONE).get_data().decode()
+    component_spec = ComponentSpec(presenter_class, template, modules, auto_bind)
+
+    template_info = TemplateInfo(template)
+    host_name = template_info.name
+    host_parent = template_info.parent
+
+    install_signals = {}
+    for name, signal in presenter_class.__install_signals__.items():
+        raise NotImplementedError
+
+    install_methods = {}
+    for name, func in presenter_class.__install_methods__.items():
+        install_methods[name] = InstalledMethod(func)
+
+    install_props = {}
+    for name, prop in presenter_class.__install_props__.items():
+        args = prop.get_pspec_args()
+        flags = args[-1]
+        flags |= GObject.ParamFlags.EXPLICIT_NOTIFY
+        install_props[name] = args[:-1] + (flags,)
+
+    namespace = {
+        **install_methods,
+
+        '__gtype_name__': host_name,
+        '__component__': component_spec,
+
+        '__gproperties__': install_props,
+        '__gsignals__': install_signals,
+
+        '__init__': _component_init,
+
+        'do_get_property': _component_get_property,
+        'do_set_property': _component_set_property,
+    }
+
+    host_type = GObjectMeta(host_name, (host_parent,), namespace)
+
+    # Not really necessary, but make it explicit that creating a new GObject subclass has a side-effect that
+    # registers the new class with the GLib type system.
+    GObject.type_register(host_type)
+
+
+class ComponentSpec:
+    def __init__(
+            self,
[…]
