+++ opendrop/app/ift/image_processing/__init__.py
@@ -1,1 +1,1 @@
-from .image_processing import IFTImageProcessing
+from . import image_processing
