+++ opendrop/app/__init__.py
@@ -1,1 +1,2 @@
+from . import main_menu, ift
 from .app import OpendropApplication
