### 2020-07-17 7e4c3a9
Move more component creation into template files
