### 2020-07-17 4505048
Add support for installing signals, and refactor out OpendropService
