+++ opendrop/app/main_menu.py
@@ -1,20 +1,22 @@
-from gi.repository import Gtk
+from gi.repository import Gtk, GObject
 
-from injector import inject
-from opendrop.appfw import component, Presenter
-from opendrop.app.services.app import OpendropService
+from opendrop.appfw import Presenter, component, install
 
 
 @component(
     template_path='./main_menu.ui',
 )
 class MainMenuPresenter(Presenter[Gtk.Window]):
-    @inject
-    def __init__(self, app_service: OpendropService) -> None:
-        self.app_service = app_service
-
     def hdl_ift_btn_clicked(self, button: Gtk.Button) -> None:
-        self.app_service.goto_ift()
+        self.ift.emit()
 
     def hdl_conan_btn_clicked(self, button: Gtk.Button) -> None:
-        self.app_service.goto_conan()
+        self.conan.emit()
+
+    @install
+    @GObject.Signal
+    def ift(self) -> None: pass
+
+    @install
+    @GObject.Signal
+    def conan(self) -> None: pass
