+++ opendrop/app/ift/services/report/report.py
@@ -1,6 +1,0 @@
-from opendrop.utility.bindable import VariableBindable
-
-
-class IFTReportService:
-    def __init__(self) -> None:
-        self.bn_analyses = VariableBindable(tuple())
