+++ opendrop/app/__init__.py
@@ -1,2 +1,2 @@
-from . import main_menu, ift
+from . import main_menu, ift, conan
 from .app import OpendropApplication
