+++ opendrop/app/conan/__init__.py
@@ -27,5 +27,4 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from .component import conan_root_cs
-from .model import ConanSession
+from . import conan_experiment

[…]
