### 2020-07-17 4cf65ae
Move image acquisition to use new appfw
