### 2020-07-17 2af8737
Refactor ift
