### 2020-07-20 d0bbf09
Move conan image processing to new appfw
