+++ opendrop/app/conan/image_processing/services/image_processing.py
@@ -27,8 +27,10 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from typing import Callable, Optional
+import asyncio
+from typing import Optional
 
+from injector import inject
 import numpy as np
 
 from opendrop.app.common.services.acquisition import ImageAcquisitionService
@@ -36,7 +38,6 @@ from opendrop.app.common.image_processing.plugins.define_line import DefineLineP
 from opendrop.app.common.image_processing.plugins.define_region import DefineRegionPluginModel
 from opendrop.app.conan.analysis import FeatureExtractor, FeatureExtractorParams, ContactAngleCalculatorParams
 from opendrop.utility.bindable import VariableBindable, AccessorBindable
-from opendrop.utility.bindable.typing import Bindable
 from opendrop.utility.geometry import Rect2
 from .plugins import ToolID
 from .plugins.foreground_detection import ForegroundDetectionPluginModel
@@ -44,17 +45,18 @@ from .plugins.preview import ConanPreviewPluginModel
 
 
 class ConanImageProcessingModel:
+    @inject
     def __init__(
             self, *,
             image_acquisition: ImageAcquisitionService,
             feature_extractor_params: FeatureExtractorParams,
             conancalc_params: ContactAngleCalculatorParams,
-            do_extract_features: Callable[[Bindable[np.ndarray], FeatureExtractorParams], FeatureExtractor]
     ) -> None:
+        self._loop = asyncio.get_event_loop()
+
         self._image_acquisition = image_acquisition
         self._feature_extractor_params = feature_extractor_params
         self._conancalc_params = conancalc_params
-        self._do_extract_features = do_extract_features
 
         self.bn_active_tool = VariableBindable(ToolID.DROP_REGION)
 
@@ -79,8 +81,11 @@ class ConanImageProcessingModel:
         self.preview_plugin = ConanPreviewPluginModel(
             image_acquisition=image_acquisition,
             feature_extractor_params=feature_extractor_params,
-            do_extract_features=do_extract_features,
+            do_extract_features=self._extract_features,
         )
+
+    def _extract_features(self, image: np.ndarray) -> FeatureExtractor:
+        return FeatureExtractor(image, self._feature_extractor_params, loop=self._loop)
 
     def _get_region_clip(self) -> Optional[Rect2[int]]:
         image_size_hint = self._image_acquisition.get_image_size_hint()
