+++ opendrop/app/conan/image_processing/__init__.py
@@ -27,5 +27,4 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from .component import conan_image_processing_cs
-from .model import ConanImageProcessingModel
+from . import image_processing

[…]
