### 2020-07-21 c46f7e1
Refactor remaining conan stuff to new appfw
Also refactor a bit off ift.. and fix some bugs there...
