+++ opendrop/app/conan/report/overview/__init__.py
@@ -27,4 +27,4 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from .component import individual_cs
+from . import overview, master, detail
