+++ opendrop/app/conan/results/component.py
@@ -1,344 +1,0 @@
-# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
-# OpenDrop is released under the GNU GPL License. You are free to
-# modify and distribute the code, but always under the same license
-# (i.e. you cannot make commercial derivatives).
-#
-# If you use this software in your research, please cite the following
-# journal articles:
-#
-# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
-# R. F. Tabor, Measurement of surface and interfacial tension using
-# pendant drop tensiometry. Journal of Colloid and Interface Science 454
-# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
-#
-# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
-# and J. D. Berry, OpenDrop: Open-source software for pendant drop
-# tensiometry & contact angle measurements, submitted to the Journal of
-# Open Source Software
-#
-# These citations help us not only to understand who is using and
-# developing OpenDrop, and for what purpose, but also to justify
-# continued development of this code and other open source resources.
-#
-# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
-# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-# PURPOSE.  See the GNU General Public License for more details.  You
-# should have received a copy of the GNU General Public License along
-# with this software.  If not, see <https://www.gnu.org/licenses/>.
-
-
-import asyncio
-import math
-from typing import Optional
-
-from gi.repository import Gtk
-
-from opendrop.app.common.footer.results import results_footer_cs, ResultsFooterStatus
-from opendrop.app.common.wizard import WizardPageControls
-from opendrop.app.conan.analysis_saver import ConanAnalysisSaverOptions, conan_save_dialog_cs
-from opendrop.app.conan.results.graphs import graphs_cs
-from opendrop.app.conan.results.individual.component import individual_cs
-from opendrop.mvp import ComponentSymbol, View, Presenter
-from opendrop.utility.bindable import VariableBindable
-from opendrop.widgets.yes_no_dialog import YesNoDialog
-from .model import ConanResultsModel
-
-conan_results_cs = ComponentSymbol()  # type: ComponentSymbol[Gtk.Widget]
-
-
-@conan_results_cs.view(options=['footer_area'])
-class ConanResultsView(View['ConanResultsPresenter', Gtk.Widget]):
-    def _do_init(self, footer_area: Gtk.Grid) -> Gtk.Widget:
-        self._widget = Gtk.Grid()
-
-        frame = Gtk.Frame(margin=10, label_xalign=0.5)
-        frame.show()
-        self._widget.attach(frame, 0, 1, 1, 1)
-
-        self._stack = Gtk.Stack()
-        self._stack.show()
-        frame.add(self._stack)
-
-        _, self._individual_area = self.new_component(
-            individual_cs.factory(
-                model=self.presenter.individual_model
-            )
-        )
-        self._individual_area.show()
-        self._stack.add_titled(self._individual_area, name='Individual Fit', title='Individual Fit')
-
-        _, self._graphs_area = self.new_component(
-            graphs_cs.factory(
-                model=self.presenter.graphs_model
-            )
-        )
-        self._graphs_area.show()
-        self._stack.add_titled(self._graphs_area, name='Graphs', title='Graphs')
-
-        self._stack_switcher = Gtk.StackSwitcher(stack=self._stack)
-        self._stack_switcher.show()
-        frame.props.label_widget = self._stack_switcher
-
-        self._stack.props.visible_child = self._individual_area
-
-        _, footer_inner = self.new_component(
-            results_footer_cs.factory(
[…]
