+++ opendrop/app/conan/report/__init__.py
@@ -27,17 +27,4 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from typing import Sequence, Optional
-
-from opendrop.app.conan.analysis import ConanAnalysis
-from opendrop.utility.bindable.typing import Bindable
-
-
-class IndividualModel:
-    def __init__(
-            self,
-            in_analyses: Bindable[Sequence[ConanAnalysis]],
-            bind_selection: Bindable[Optional[ConanAnalysis]]
-    ) -> None:
-        self.bn_analyses = in_analyses
-        self.bn_selection = bind_selection
+from . import report, overview, graphs

[…]
