### 2020-07-21 71963c7
Remove common/wizard stuff
