### 2020-07-21 5ab67ae
Refactor linear footer to new appfw
