+++ opendrop/app/conan/conan_experiment.py
@@ -30,7 +30,6 @@
 from gi.repository import GObject, Gtk
 from injector import inject
 
-from opendrop.app.common.footer.linearnav import linear_navigator_footer_cs
 from opendrop.app.common.footer.results import results_footer_cs, ResultsFooterStatus
 from opendrop.utility.bindable import VariableBindable
 from opendrop.appfw import ComponentFactory, Presenter, TemplateChild, component
@@ -47,8 +46,6 @@ from .services.session import ConanSession, ConanSessionModule
 )
 class ConanExperimentPresenter(Presenter[Gtk.Assistant]):
     action_area = TemplateChild('action_area')  # type: TemplateChild[Gtk.Stack]
-    action0 = TemplateChild('action0')  # type: TemplateChild[Gtk.Container]
-    action1 = TemplateChild('action1')  # type: TemplateChild[Gtk.Container]
     action2 = TemplateChild('action2')  # type: TemplateChild[Gtk.Container]
 
     report_page = TemplateChild('report_page')
@@ -67,13 +64,6 @@ class ConanExperimentPresenter(Presenter[Gtk.Assistant]):
         session.bind_property('analyses', self.progress_helper, 'analyses', GObject.BindingFlags.SYNC_CREATE)
 
     def after_view_init(self) -> None:
-        self.lin_footer_component = linear_navigator_footer_cs.factory(
-            do_back=self.previous_page,
-            do_next=self.next_page,
-        ).create()
-        self.lin_footer_component.view_rep.show()
-        self.action0.add(self.lin_footer_component.view_rep)
-
         self._results_footer_status = VariableBindable(ResultsFooterStatus.IN_PROGRESS)
         self._results_footer_progress = VariableBindable(0.0)
         self._results_footer_time_elapsed = VariableBindable(0.0)
@@ -123,12 +113,9 @@ class ConanExperimentPresenter(Presenter[Gtk.Assistant]):
 
     def prepare(self, *_) -> None:
         cur_page = self.host.get_current_page()
-        if cur_page in (0, 1):
-            self.action_area.set_visible_child_name('0')
-        else:
-            self.action_area.set_visible_child_name(str(cur_page))
+        self.action_area.set_visible_child_name(str(cur_page))
 
-    def next_page(self) -> None:
+    def next_page(self, *_) -> None:
         cur_page = self.host.get_current_page()
         if cur_page == 0:
             self.host.next_page()
@@ -139,7 +126,7 @@ class ConanExperimentPresenter(Presenter[Gtk.Assistant]):
             # Ignore, on last page.
             return
 
-    def previous_page(self) -> None:
+    def previous_page(self, *_) -> None:
         cur_page = self.host.get_current_page()
         if cur_page == 0:
             # Ignore, on first page.
@@ -234,5 +221,4 @@ class ConanExperimentPresenter(Presenter[Gtk.Assistant]):
         return True
 
     def destroy(self, *_) -> None:
-        self.lin_footer_component.destroy()
         self.results_footer_component.destroy()
