### 2020-07-21 c48a24a
Refactor results footer to use new appfw
