+++ opendrop/app/common/footer/analysis/_types.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class AnalysisFooterStatus(Enum):
+        IN_PROGRESS = 0
+        FINISHED = 1
+        CANCELLED = 2
