+++ opendrop/app/common/footer/analysis/__init__.py
@@ -27,10 +27,5 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from enum import Enum
-
-
-class ResultsFooterStatus(Enum):
-        IN_PROGRESS = 0
-        FINISHED = 1
-        CANCELLED = 2
+from . import _analysis, _bar
+from ._types import AnalysisFooterStatus

[…]
