### 2020-07-21 6f9d61e
Add version info to main menu
