### 2020-07-21 c7c2190
Fix progress bar time elapsed not updating
