### 2020-07-21 fef33b2
Update docs url
Remove /en/latest/ from the end of the link to not be specific about the docs version.
