### 2020-07-21 41a053a
Improve formatting
Un-bold second mention of "OpenDrop", having the first one in bold is enough I think.
