### 2020-07-23 7ea18af
Patch typing module with backports
Instead of importing typing_extensions, which is not necessary for
new enough python versions.
