### 2020-07-23 b76070e
Add width and height attributes to Rect2
They are the same as w and h.
