### 2020-07-23 637d392
Edit error message when invalid arguments passed to constructors
