### 2020-07-27 6716e53
Add initial files for new Canvas widget
To replace the current Render widget
