+++ opendrop/widgets/canvas/__init__.py
@@ -1,0 +1,3 @@
+from ._canvas import *
+from ._artist import *
+from ._affine import *
