+++ opendrop/widgets/canvas/_canvas.py
@@ -1,0 +1,271 @@
+from enum import Enum
+from typing import Optional, MutableMapping, Tuple
+
+import cairo
+from gi.repository import GObject, Gdk, Gtk
+
+from opendrop.utility.geometry import Rect2
+
+from ._affine import AffineTransformArtist
+from ._artist import Artist, ArtistContainer
+
+
+__all__ = ('Canvas', 'CanvasAlign')
+
+
+class CanvasAlign(Enum):
+    FIT = 1
+    CENTER = 2
+
+
+# add surface? here, CANVAS, CANVAS_SCALED, VIEWPORT
+# zoom(request, around)
+# get_zoom
+# get_zoom_request
+# ::zoom signal
+
+class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
+    _zoom = 1
+    _content_area = Rect2(0, 0, 0, 0)
+
+    _align = CanvasAlign.FIT
+
+    _hscroll_policy = Gtk.ScrollablePolicy.MINIMUM
+    _vscroll_policy = Gtk.ScrollablePolicy.MINIMUM
+
+    def __init__(self, **properties) -> None:
+        self._artist = AffineTransformArtist()
+        self._artist.set_invalidate_handler(self._artist_invalidated)
+
+        self._adjustments = {
+            Gtk.Orientation.HORIZONTAL: None,
+            Gtk.Orientation.VERTICAL: None,
+        }  # type: MutableMapping[Gtk.Orientation, Optional[Gtk.Adjustment]]
+
+        self._adjustment_handler_ids = {
+            Gtk.Orientation.HORIZONTAL: None,
+            Gtk.Orientation.VERTICAL: None,
+        }  # type: MutableMapping[Gtk.Orientation, Optional[int]]
+
+        super().__init__(**properties)
+
+        self._default_adjustments_if_none()
+
+    def _get_adjustment(self, orientation: Gtk.Orientation) -> Gtk.Adjustment:
+        adjustment = self._adjustments[orientation]
+        if adjustment is None:
+            raise AttributeError
+        return adjustment
+
+    def _set_adjustment(
+            self,
+            orientation: Gtk.Orientation,
+            adjustment: Optional[Gtk.Adjustment] = None
+    ) -> None:
+        if adjustment and self._get_adjustment(orientation) == adjustment:
+            return
+
+        self._disconnect_adjustment(orientation)
+
+        adjustment = adjustment or Gtk.Adjustment()
+        self._adjustments[orientation] = adjustment
+        self._configure_adjustment(orientation)
+
+        self._adjustment_handler_ids[orientation] = \
+            adjustment.connect('value-changed', self._adjustment_value_changed)
+
+        self._adjustment_value_changed(adjustment)
+
+    def _configure_adjustment(self, orientation: Gtk.Orientation) -> None:
+        if orientation is Gtk.Orientation.HORIZONTAL:
+            viewport_size = self.get_allocation().width
+            content_size = self._content_area.width
+        else:
+            viewport_size = self.get_allocation().height
+            content_size = self._content_area.height
+
+        adjustment = self._get_adjustment(orientation)
+
+        adjustment.configure(
+            adjustment.get_value(),  # value
+            0,  # lower
+            max(self._zoom * content_size, viewport_size),  # upper
+            viewport_size * 0.1,  # step_increment
+            viewport_size * 0.9,  # page_increment
+            viewport_size,  # page_size
+        )
+
+    def _disconnect_adjustment(self, orientation: Gtk.Orientation) -> None:
+        adjustment = self._adjustments[orientation]
+        if adjustment is None: return
+
[…]

[…]
