### 2020-07-27 92e8f5d
Rename image acquirer type names
Rename 'Local filesystem' to 'Filesystem'.
Rename 'USB Camera' to 'OpenCV'.
