### 2020-07-27 da8f526
Block GObject.Object initialisation until later
