### 2020-07-27 0e84bbf
Make IntegerEntry more GtkBuilder friendly
Same as was done for FloatEntry in a previous commit.
Also remove some unnecessary code from FloatEntry.
