### 2020-07-28 882a2ae
Support CONSTRUCT_ONLY presenter properties
