### 2020-07-28 d57a507
Add support for GenICam compliant cameras
