+++ opendrop/app/common/image_acquisition/configurator/__init__.py
@@ -27,4 +27,4 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from .component import configurator_cs
+from . import configurator, genicam
