+++ opendrop/app/common/services/acquisition/__init__.py
@@ -1,2 +1,2 @@
 from ._acquisition import ImageAcquisitionService, AcquirerType
-from ._acquirer import ImageAcquirer, InputImage, ImageSequenceAcquirer, CameraAcquirer, LocalStorageAcquirer, USBCameraAcquirer
+from ._acquirer import ImageAcquirer, InputImage, ImageSequenceAcquirer, CameraAcquirer, LocalStorageAcquirer, USBCameraAcquirer, GenicamAcquirer
