+++ opendrop/app/common/services/acquisition/_acquirer/__init__.py
@@ -32,3 +32,4 @@ from .camera import CameraAcquirer
 from .image_sequence import ImageSequenceAcquirer
 from .local_storage import LocalStorageAcquirer
 from .usb_camera import USBCameraAcquirer
+from .genicam import GenicamAcquirer

[…]
