### 2020-07-28 3b76f9a
Add genicam to requirements
Needed by harvesters.
