### 2020-07-28 88003dc
Remove enscons, use simple shim instead
Wasn't using enscons correctly (forgot to specify build-backend but was
still getting errors afterwards anyway) and didn't want to figure out
how to.
