### 2020-08-24 d61841e
Fix incorrect scaling when downsampling pixel bits
