### 2020-08-24 775e543
Use os.pathsep to determine PATH-like path separators
Currently wrongly assuming it is ';'.
