### 2020-08-24 e5cdfd2
Hide GenicamAcquirer details and handle not implemented device features
Mainly add exception handling for when a device does not implement
'DEVICE_INFO_USER_DEFINED_NAME' and 'DEVICE_INFO_VERSION' values and
provide fallback defaults (instead of crashing oops).
