### 2020-08-27 #2128 @kyleniemeyer
Hi @jdber1, sorry for letting this slide. Part of the problem is none of the folks I emailed to review ever responded... additional suggestions would be welcome, but I will also seek out another from the JOSS reviewer pool (we need two reviewers).
