### 2020-08-27 #2604 @whedon
[REVIEW]: OpenDrop: Open-source software for pendant drop tensiometry & contact angle measurements
**Submitting author:** @jdber1 (<a href="http://orcid.org/0000-0002-0961-7782">Joe Berry</a>)
**Repository:** <a href="https://github.com/jdber1/opendrop" target ="_blank">https://github.com/jdber1/opendrop</a>
**Version:** v3.2.2
