### 2020-08-27 #2604 @kyleniemeyer
@jdber1 while the reviewers begin looking at the software, there are a few issues with the paper you can take care of. It looks like your bib file has a number of references that are not cited in the paper, and then any references you do cite need DOIs
