### 2020-08-31 82fcfcd
Update paper.bib
Deleted superfluous references, added dois.
