+++ paper/paper.bib
@@ -1,110 +1,3 @@
-@article{Bremond2008,
-abstract = {The destabilization process of an emulsion under flow is investigated in a microfluidic device. The experimental approach enables us to generate a periodic train of droplet pairs, and thus to isolate and analyze the basic step of the destabilization, namely, the coalescence of two droplets which collide. We demonstrate a counterintuitive phenomenon: coalescence occurs during the separation phase and not during the impact. Separation induces the formation of two facing nipples in the contact area that hastens the connection of the interfaces prior to fusion. Moreover, droplet pairs initially stabilized by surfactants can be destabilized by forcing the separation. Finally, we note that the fusion mechanism is responsible for a cascade of coalescence events in a compact system of droplets where the separation is driven by surface tension.},
-author = {Bremond, Nicolas and Thiam, Abdou R. and Bibette, J{\'{e}}r{\^{o}}me},
-file = {:Users/berryj/ownCloud/Papers/Mendeley/Bremond, Thiam, Bibette{\_}2008{\_}Physical Review Letters.pdf:pdf},
-isbn = {0031-9007},
-issn = {00319007},
-journal = {Physical Review Letters},
-mendeley-groups = {Dagastine/Coalescence},
-number = {2},
-pages = {1--4},
-pmid = {18232876},
-title = {{Decompressing emulsion droplets favors coalescence}},
-volume = {100},
-year = {2008}
-}
-
-
-@article{Eggers1997,
-annote = {NULL},
-author = {Eggers, Jens},
-journal = {Review of Modern Physics},
-mendeley-groups = {Rico/pendant drop,unimelb,unimelb/electrokinetics},
-number = {3},
-pages = {865--929},
-title = {{Nonlinear dynamics and breakup of free-surface flows}},
-volume = {69},
-year = {1997}
-}
-
-@article{Eggers2008,
-abstract = {Jets, i.e. collimated streams of matter, occur from the microscale up to the large-scale structure of the universe. Our focus will be mostly on surface tension effects, which result from the cohesive properties of liquids. Paradoxically, cohesive forces promote the breakup of jets, widely encountered in nature, technology and basic science, for example in nuclear fission, DNA sampling, medical diagnostics, sprays, agricultural irrigation and jet engine technology. Liquid jets thus serve as a paradigm for free-surface motion, hydrodynamic instability and singularity formation leading to drop breakup. In addition to their practical usefulness, jets are an ideal probe for liquid properties, such as surface tension, viscosity or non-Newtonian rheology. They also arise from the last but one topology change of liquid masses bursting into sprays. Jet dynamics are sensitive to the turbulent or thermal excitation of the fluid, as well as to the surrounding gas or fluid medium. The aim of this review is to provide a unified description of the fundamental and the technological aspects of these subjects.},
-author = {Eggers, Jens and Villermaux, Emmanuel},
-file = {::},
-isbn = {0034-4885},
-issn = {0034-4885},
-journal = {Reports on Progress in Physics},
-mendeley-groups = {Rico/pendant drop},
-pages = {036601},
-title = {{Physics of liquid jets}},
-volume = {71},
-year = {2008}
-}
-
-@article{Good1992,
-author = {Good, Robert J.},
-journal = {Journal of Adhesion Science and Technology},
-pages = {1269--1302},
-pmid = 25246403,
[…]
