### 2020-08-31 #2604 @jdber1
> @jdber1 while the reviewers begin looking at the software, there are a few issues with the paper you can take care of. It looks like your bib file has a number of references that are not cited in the paper, and then any references you do cite need DOIs

@kyleniemeyer I have updated the bib file to include DOIs and also delete any un-cited references.
