### 2020-09-07 jdber1/opendrop#23 @rajeshrinet
Community guidelines
@jdber1 , JOSS requires **Community guidelines** for

1) Contribute to the software. *This is best achieved by adding a file as described [here](https://docs.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors).*

2) Report issues or problems with the software

3) Seek support

(2) and (3) can be described in the README.
