### 2020-09-08 059a2b7
Create CONTRIBUTING.md
