### 2020-09-08 29a1703
Update app starting procedure
Starter script in PATH has been removed.
