### 2020-09-08 7db2b3f
Fix Python 3.5 compatibility when using Generic with GObject
Add a syntax stub to mimic Generic.
