### 2020-09-08 #2604 @rajeshrinet
@ kyleniemeyer I have now gone through the paper and successfully installed the library from the GitHub repo. The paper is well-written and the sofware looks very good! Please can I request the authors to respond to my following comments such that I can complete my review checklist:

1. The **installation** works as indicated on the [docs](https://opendrop.readthedocs.io/en/latest/getting_started/index.html) but the command for launch needs update: `python3 -m opendrop`

2. **Example usage**: It would be useful to provide steps to run examples from sample frames on the repo. This can then also be used to test as described in 3 below.

3. **Testing:** The library does have a `tests` folder. But there does not seem to be an automated testing nor there is an explanation of manual steps for testing.

4. **Community guidelines** needs to be added to the repo as indicate in this issue: https://github.com/jdber1/opendrop/issues/23

5. **State of the field**: The author should compare opendrop with existing similar software, if any.
