### 2020-09-11 0163b4b
Update paper.md
Adding comparison figures
