+++ paper/paper.md
@@ -70,6 +70,8 @@ The experimental apparatus required for measurements of interfacial tension and 
 
 Here we present the latest version of OpenDrop. The new version, Barracuda, is able to measure interfacial tension and also contact angle in a variety of configurations with field-leading accuracy and reproducibility. The contact angle measurement capability is new for this release, but has been used successfully in previous studies [@Prathapan2017]. OpenDrop has been written in Python because it is open-source, free, runs on multiple operating systems (including Linux, Mac OSX and Windows), and is easily integrable with a large number of mature, 3rd party open source libraries. In particular, OpenDrop utilises the sophisticated image processing capabilities of the OpenCV library in order to extract drop profiles from experimental images for input into the requisite numerical algorithm. Further, the ease of readability and modular nature of Python encourages and supports collaboration, and gives OpenDrop significant pedagogic value. Python can also be easily integrated with other languages, of particular importance to pendant drop tensiometry and contact angle measurements where integration of code needed to control cameras and associated software is a critical requirement. The previous version is in use in many research groups around the world, and is also used in teaching laboratories including Monash University. 
 
+![Caption for example figure.\label{fig:example}](iftFigure.pdf)
+
 
 
 The availability of the software allows the interested user to
