### 2020-09-11 0e7266b
Added contact angle figure
