### 2020-10-06 c45ab86
Fix bug, cannot dictionary unpack SCons.Environment
