### 2020-10-06 c59a152
Start work on PendantEdgeDetectionService
Intended to replace FeatureExtractorService.
