+++ opendrop/app/ift/services/edges.py
@@ -1,0 +1,164 @@
+# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
+# OpenDrop is released under the GNU GPL License. You are free to
+# modify and distribute the code, but always under the same license
+# (i.e. you cannot make commercial derivatives).
+#
+# If you use this software in your research, please cite the following
+# journal articles:
+#
+# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
+# R. F. Tabor, Measurement of surface and interfacial tension using
+# pendant drop tensiometry. Journal of Colloid and Interface Science 454
+# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
+#
+# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
+# and J. D. Berry, OpenDrop: Open-source software for pendant drop
+# tensiometry & contact angle measurements, submitted to the Journal of
+# Open Source Software
+#
+# These citations help us not only to understand who is using and
+# developing OpenDrop, and for what purpose, but also to justify
+# continued development of this code and other open source resources.
+#
+# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
+# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+# PURPOSE.  See the GNU General Public License for more details.  You
+# should have received a copy of the GNU General Public License along
+# with this software.  If not, see <https://www.gnu.org/licenses/>.
+
+
+from opendrop.utility.geometry import Rect2
+from typing import Optional
+
+from gi.repository import GObject
+
+import numpy as np
+
+from opendrop.processing.ift import (
+    apply_edge_detection,
+    extract_drop_profile,
+    extract_needle_profile,
+)
+
+
+class PendantEdgeDetectionParamsFactory(GObject.Object):
+    _canny_min: int = 30
+    _canny_max: int = 60
+    _needle_region: Optional[Rect2[int]] = None
+    _drop_region: Optional[Rect2[int]] = None
+
+    def create(self) -> 'PendantEdgeDetectionParams':
+        return PendantEdgeDetectionParams(
+            canny_min=self._canny_min,
+            canny_max=self._canny_max,
+            needle_region=self._needle_region,
+            drop_region=self._drop_region,
+        )
+
+    @GObject.Signal
+    def changed(self) -> None:
+        """Emitted when edge detection parameters are changed."""
+
+    @GObject.Property
+    def needle_region(self) -> Optional[Rect2[int]]:
+        return self._needle_region
+
+    @needle_region.setter
+    def needle_region(self, region: Optional[Rect2[int]]) -> None:
+        self._needle_region = region
+        self.changed.emit()
+
+    @GObject.Property
+    def drop_region(self) -> Optional[Rect2[int]]:
+        return self._drop_region
+
+    @drop_region.setter
+    def drop_region(self, region: Optional[Rect2[int]]) -> None:
+        self._drop_region = region
+        self.changed.emit()
+
+    @GObject.Property
+    def canny_min(self) -> int:
+        return self._canny_min
+
+    @canny_min.setter
+    def canny_min(self, value: int) -> None:
+        self._canny_min = value
+        self.changed.emit()
+
+    @GObject.Property
+    def canny_max(self) -> int:
+        return self._canny_max
+
+    @canny_max.setter
+    def canny_max(self, value: int) -> None:
+        self._canny_max = value
+        self.changed.emit()
+
+
+class PendantEdgeDetectionParams:
+    """Plain Old Data structure"""
+
+    canny_min: int
+    canny_max: int
+    needle_region: Optional[Rect2[int]]
[…]
