### 2020-10-06 ef9b389
Rename AffineTransformArtist to AffineFrameArtist
