### 2020-10-06 83d8df3
Implement Canvas zooming
