+++ opendrop/widgets/canvas/_canvas.py
@@ -18,17 +18,20 @@ class CanvasAlign(Enum):
     CENTER = 2
 
 
-# add surface? here, CANVAS, CANVAS_SCALED, VIEWPORT
-# zoom(request, around)
-# get_zoom
-# get_zoom_request
+# add surface? here, CANVAS, BIN, VIEWPORT
 # ::zoom signal
 
+# TODO: handle when content_area is not at origin
+
 class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
-    _zoom = 1
+    _scale_request = 1
+    _scale = 1
+    _viewport_origin = (0, 0)
     _content_area = Rect2(0, 0, 0, 0)
 
     _align = CanvasAlign.FIT
+
+    _ignore_adjustment_changes = False
 
     _hscroll_policy = Gtk.ScrollablePolicy.MINIMUM
     _vscroll_policy = Gtk.ScrollablePolicy.MINIMUM
@@ -67,32 +70,42 @@ class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
 
         self._disconnect_adjustment(orientation)
 
-        adjustment = adjustment or Gtk.Adjustment()
+        new_adjustment = adjustment is None
+        if new_adjustment:
+            adjustment = Gtk.Adjustment()
+
         self._adjustments[orientation] = adjustment
         self._configure_adjustment(orientation)
 
-        self._adjustment_handler_ids[orientation] = \
-            adjustment.connect('value-changed', self._adjustment_value_changed)
+        self._adjustment_handler_ids[orientation] \
+            = adjustment.connect('value-changed', self._adjustment_value_changed)
 
-        self._adjustment_value_changed(adjustment)
+        if not new_adjustment:
+            self._adjustment_value_changed(adjustment)
 
     def _configure_adjustment(self, orientation: Gtk.Orientation) -> None:
         if orientation is Gtk.Orientation.HORIZONTAL:
+            viewport_origin = self._viewport_origin[0]
             viewport_size = self.get_allocation().width
             content_size = self._content_area.width
         else:
+            viewport_origin = self._viewport_origin[1]
             viewport_size = self.get_allocation().height
             content_size = self._content_area.height
 
         adjustment = self._get_adjustment(orientation)
 
+        page_size = int(min(viewport_size, content_size * self._scale))
+        lower = 0
+        upper = int(content_size * self._scale)
+
         adjustment.configure(
-            adjustment.get_value(),  # value
-            0,  # lower
-            max(self._zoom * content_size, viewport_size),  # upper
-            viewport_size * 0.1,  # step_increment
-            viewport_size * 0.9,  # page_increment
-            viewport_size,  # page_size
+            viewport_origin,
+            lower,
+            upper,
+            page_size * 0.1,  # step_increment
+            page_size * 0.9,  # page_increment
+            page_size,  # page_size
         )
 
     def _disconnect_adjustment(self, orientation: Gtk.Orientation) -> None:
@@ -111,19 +124,38 @@ class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
                 self._set_adjustment(orientation)
 
     def _adjustment_value_changed(self, adjustment: Gtk.Adjustment) -> None:
+        if self._ignore_adjustment_changes: return
+
+        self._viewport_origin = (
+            self.hadjustment.props.value,
+            self.vadjustment.props.value,
+        )
         self.queue_allocate()
 
     def set_content_area(self, x: int, y: int, width: int, height: int) -> None:
         self._content_area = Rect2(x=x, y=y, w=width, h=height)
         self.queue_resize()
 
[…]
