### 2020-10-07 152648d
Refactor and change set_content_area() to set_content_size()
Don't allow arbitrary content regions, just use width and height for simplicity.
