+++ opendrop/widgets/canvas/_canvas.py
@@ -4,8 +4,6 @@ from typing import Optional, MutableMapping, Tuple
 import cairo
 from gi.repository import GObject, Gdk, Gtk
 
-from opendrop.utility.geometry import Rect2
-
 from ._affine import AffineFrameArtist
 from ._artist import Artist, ArtistContainer
 
@@ -18,23 +16,23 @@ class CanvasAlign(Enum):
     CENTER = 2
 
 
-# add surface? here, CANVAS, BIN, VIEWPORT
+# surface type: CANVAS, CHILD, VIEWPORT
 # ::zoom signal
 
-# TODO: handle when content_area is not at origin
-
 class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
-    _scale_request = 1
+    _MIN_SCALE = 0.01
+    _MAX_SCALE = 100
+
+    _viewport_position = (0, 0)
+
+    _content_size = (0, 0)
     _scale = 1
-    _viewport_origin = (0, 0)
-    _content_area = Rect2(0, 0, 0, 0)
-
     _align = CanvasAlign.FIT
-
-    _ignore_adjustment_changes = False
 
     _hscroll_policy = Gtk.ScrollablePolicy.MINIMUM
     _vscroll_policy = Gtk.ScrollablePolicy.MINIMUM
+
+    _ignore_adjustment_changes = False
 
     def __init__(self, **properties) -> None:
         self._artist = AffineFrameArtist()
@@ -65,8 +63,12 @@ class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
             orientation: Gtk.Orientation,
             adjustment: Optional[Gtk.Adjustment] = None
     ) -> None:
-        if adjustment and self._get_adjustment(orientation) == adjustment:
-            return
+        if adjustment is not None:
+            try:
+                if self._get_adjustment(orientation) == adjustment:
+                    return
+            except AttributeError:
+                pass
 
         self._disconnect_adjustment(orientation)
 
@@ -85,22 +87,22 @@ class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
 
     def _configure_adjustment(self, orientation: Gtk.Orientation) -> None:
         if orientation is Gtk.Orientation.HORIZONTAL:
-            viewport_origin = self._viewport_origin[0]
             viewport_size = self.get_allocation().width
-            content_size = self._content_area.width
+            viewport_position = self._viewport_position[0]
+            content_size = self._content_size[0]
         else:
-            viewport_origin = self._viewport_origin[1]
             viewport_size = self.get_allocation().height
-            content_size = self._content_area.height
+            viewport_position = self._viewport_position[1]
+            content_size = self._content_size[1]
 
         adjustment = self._get_adjustment(orientation)
 
-        page_size = int(min(viewport_size, content_size * self._scale))
         lower = 0
-        upper = int(content_size * self._scale)
+        upper = int(self._scale * content_size)
+        page_size = min(viewport_size, upper - lower)
 
         adjustment.configure(
-            viewport_origin,
+            viewport_position,
             lower,
             upper,
             page_size * 0.1,  # step_increment
@@ -126,36 +128,31 @@ class Canvas(Gtk.DrawingArea, Gtk.Scrollable, ArtistContainer):
     def _adjustment_value_changed(self, adjustment: Gtk.Adjustment) -> None:
         if self._ignore_adjustment_changes: return
 
-        self._viewport_origin = (
-            self.hadjustment.props.value,
-            self.vadjustment.props.value,
-        )
+        self._viewport_position = (self.hadjustment.props.value, self.vadjustment.props.value)
         self.queue_allocate()
 
[…]
