### 2020-10-07 d03b68a
Update index.rst
Updated launch command
