### 2020-10-07 c66560c
Update README.rst
Added link to sample images and instructions for processing, and support contact details
