+++ README.rst
@@ -11,6 +11,22 @@ It runs on **Python 3.5+**, and uses **GTK+ 3** for the user interface, and **Op
 
 OpenDrop is released under the **GNU GPL** open source license, and available for free.
 
-----
+*********
+Installation & User Guide
+*********
+**OpenDrop** currently runs on Ubuntu, MacOS and Windows 10 operating systems.
+For installation instructions and user guide, visit: https://opendrop.readthedocs.io/
 
-For installation instructions and user guides, visit: https://opendrop.readthedocs.io/
+*********
+Examples
+*********
+
+Example images are located in the example_images subfolder. Instructions on how to process the pendant drop image to extract surface tension, and the images of drops on surfaces to extract contact angle, are also available at: https://opendrop.readthedocs.io/
+
+*********
+Support
+*********
+For any questions, issues, or feedback feel free to open an issue here: https://github.com/jdber1/opendrop/issues
+
+We can also be contacted by email `here <mailto:opendrop.dev@gmail.com>`_
+
