### 2020-10-07 a298a89
Update paper.bib
Added contact angle references
