+++ paper/paper.bib
@@ -20,6 +20,41 @@ file = {:Users/berryj/ownCloud/Papers/Mendeley/Prathapan et al.{\_}2017{\_}ACS A
 journal = {ACS Applied Materials {\&} Interfaces},
 keywords = {alignment,cellulose nanocrystals,hydrophobicity,pinning,wettability},
 title = {{Decreasing the Wettability of Cellulose Nanocrystal Surfaces Using Wrinkle-Based Alignment}},
+pages = {15202−-15211},
 doi = {10.1021/acsami.7b03094}, 
 year = {2017}
 }
+
+@article{Brown2016,
+author = {Brown, Philip S and Bhushan, Bharat},
+journal = {Scientific Reports},
+pages = {21048},
+title = {{Durable, superoleophobic polymer – nanoparticle composite surfaces with re-entrant geometry via solvent-induced phase transformation}},
+doi = {10.1038/srep21048},
+volume = {6},
+year = {2016}
+}
+
+@article{Nie2017,
+author = {Nie, Bin'en and Long, Teng and Li, Hui and Wang, Xiaojie and Yue, Bing},
+doi = {10.1039/C7RA05538B},
+journal = {RSC Advances},
+mendeley-groups = {Rico/pendant drop},
+pages = {34321--34330},
+title = {{A comparative analysis of antibacterial properties and in flammatory responses for the KR-12 peptide on titanium and PEGylated titanium surfaces}},
+volume = {7},
+year = {2017}
+}
+
+@phdthesis{Stacy2009,
+author = {Stacy, Russell},
+booktitle = {M. S. Thesis},
+file = {:Users/berryj/Dropbox/Papers/Mendeley/Stacy - 2009 - Master's Thesis.pdf:pdf},
+mendeley-groups = {Rico},
+pages = {83 pp},
+title = {{Contact Angle Measuremente Technique for Rough Surfaces}},
+year = {2009}
+}
+
+
+
