+++ paper/paper.md
@@ -68,10 +68,11 @@ optimization of processes involving multiphase systems.
 
 The experimental apparatus required for measurements of interfacial tension and contact angle is conceptually extremely simple, requiring only a needle, a camera, and a light source. The complexity (and associated cost of commercial instruments) comes from the image processing and the complicated numerical algorithm required to calculate these quantities from the acquired experimental image. In 2015, we released OpenDrop, which enables interfacial tension measurements more rapidly, cheaply and accurately than commercial options [@Berry2015]. The only cost to the user is the camera required (approx. $20 - $1K depending upon application), whereas commercial instruments are much more expensive (~$50K). 
 
-Here we present the latest version of OpenDrop. The new version, Barracuda, is able to measure interfacial tension and also contact angle in a variety of configurations with field-leading accuracy and reproducibility. The contact angle measurement capability is new for this release, but has been used successfully in previous studies [@Prathapan2017]. OpenDrop has been written in Python because it is open-source, free, runs on multiple operating systems (including Linux, Mac OSX and Windows), and is easily integrable with a large number of mature, 3rd party open source libraries. In particular, OpenDrop utilises the sophisticated image processing capabilities of the OpenCV library in order to extract drop profiles from experimental images for input into the requisite numerical algorithm. Further, the ease of readability and modular nature of Python encourages and supports collaboration, and gives OpenDrop significant pedagogic value. Python can also be easily integrated with other languages, of particular importance to pendant drop tensiometry and contact angle measurements where integration of code needed to control cameras and associated software is a critical requirement. The previous version is in use in many research groups around the world, and is also used in teaching laboratories including Monash University. 
+Here we present the latest version of OpenDrop. The new version, Barracuda, is able to measure interfacial tension and also contact angle in a variety of configurations with field-leading accuracy and reproducibility. The performance of OpenDrop compared to currently available commercial instrumentation is shown in \autoref{fig:ift} and \autoref{fig:ca}.
+The contact angle measurement capability is new for this release, but has been used successfully in previous studies [@Prathapan2017]. OpenDrop has been written in Python because it is open-source, free, runs on multiple operating systems (including Linux, Mac OSX and Windows), and is easily integrable with a large number of mature, 3rd party open source libraries. In particular, OpenDrop utilises the sophisticated image processing capabilities of the OpenCV library in order to extract drop profiles from experimental images for input into the requisite numerical algorithm. Further, the ease of readability and modular nature of Python encourages and supports collaboration, and gives OpenDrop significant pedagogic value. Python can also be easily integrated with other languages, of particular importance to pendant drop tensiometry and contact angle measurements where integration of code needed to control cameras and associated software is a critical requirement. The previous version is in use in many research groups around the world, and is also used in teaching laboratories including Monash University. 
 
-![Caption for example figure.\label{fig:example}](iftFigure.pdf)
-![Caption for example figure.\label{fig:example}](conAnFigure.pdf)
+![Comparison of the surface or interfacial tension of different systems calculated with OpenDrop against values reported in the literature.\label{fig:ift}](iftFigure.pdf)
+![Comparison of contact angles calculated from experimental images in the literature against values calculated with commercial instrumentation.The images are taken from [@Nie2017], [@Stacy2009] and [@Brown2016]. \label{fig:ca}](conAnFigure.pdf)
 
 
 
