### 2020-10-08 #2604 @jdber1
@rajeshrinet Thank you for your feedback. Here are the changes we have made based on your list:

> 1. The **installation** works as indicated on the [docs](https://opendrop.readthedocs.io/en/latest/getting_started/index.html) but the command for launch needs update: `python3 -m opendrop`

The [docs](https://opendrop.readthedocs.io/en/latest/getting_started/index.html) has been updated to include the correct command.

> 2. **Example usage**: It would be useful to provide steps to run examples from sample frames on the repo. This can then also be used to test as described in 3 below.

I have included sample images for both surface tension and contact angle functionality in the [example_images](https://github.com/jdber1/opendrop/tree/master/example_images) folder. The steps involved have been outlined in the documentation, see here for [interfacial tension](https://opendrop.readthedocs.io/en/latest/usage/index.html#interfacial-tension), and here for [contact angle](https://opendrop.readthedocs.io/en/latest/usage/index.html#contact-angle). I have updated the readme to include information on this.

> 3. **Testing:** The library does have a `tests` folder. But there does not seem to be an automated testing nor there is an explanation of manual steps for testing.

Manual steps for testing with the example images are outlined in the [docs](https://opendrop.readthedocs.io/en/latest/getting_started/index.html) , as outlined in the response to 2.

> 4. **Community guidelines** needs to be added to the repo as indicate in this issue: [jdber1/opendrop#23](https://github.com/jdber1/opendrop/issues/23)

Community guidelines have been added to the repo.

> 5. **State of the field**: The author should compare opendrop with existing similar software, if any.

I have modified the paper to include two figures showing comparisons between OpenDrop and existing commercial software. I am still trying to get the captions and the figure references to show correctly.
