### 2020-10-10 #2604 @rajeshrinet
@jdber1, thanks for your responses. @kyleniemeyer, I have now completed my checklists and I am happy to recommend this paper.
