### 2020-11-03 ab8d7aa
Fix bug where image navigation buttons are invisible
