### 2020-11-24 #2604 @jdber1
Hi @kyleniemeyer, do you need more reviewers to complete this review? Additionally, I am having trouble defining figure sizes and captions in the paper. Are there guidelines for this?
