### 2020-11-27 f950ce8
Add async_detect() and fix drop and needle edge detection bug
