### 2020-12-03 e6f57b7
Sizeable refactor of IFT
Computationally intensive tasks offloaded to separate processes instead
of separate threads in the same process. This should keep the UI
responsive. Also when analysing many images, their working threads won't
all compete for attention stalling the entire app.
