+++ opendrop/app/ift/services/features.py
@@ -1,207 +1,0 @@
-# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
-# OpenDrop is released under the GNU GPL License. You are free to
-# modify and distribute the code, but always under the same license
-# (i.e. you cannot make commercial derivatives).
-#
-# If you use this software in your research, please cite the following
-# journal articles:
-#
-# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
-# R. F. Tabor, Measurement of surface and interfacial tension using
-# pendant drop tensiometry. Journal of Colloid and Interface Science 454
-# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
-#
-# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
-# and J. D. Berry, OpenDrop: Open-source software for pendant drop
-# tensiometry & contact angle measurements, submitted to the Journal of
-# Open Source Software
-#
-# These citations help us not only to understand who is using and
-# developing OpenDrop, and for what purpose, but also to justify
-# continued development of this code and other open source resources.
-#
-# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
-# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-# PURPOSE.  See the GNU General Public License for more details.  You
-# should have received a copy of the GNU General Public License along
-# with this software.  If not, see <https://www.gnu.org/licenses/>.
-
-
-import asyncio
-import math
-from typing import Optional, Tuple
-
-import numpy as np
-
-from opendrop.processing.ift import (
-    apply_edge_detection,
-    extract_drop_profile,
-    extract_needle_profile,
-    calculate_width_from_needle_profile,
-)
-from opendrop.utility.bindable import VariableBindable, AccessorBindable, thread_safe_bindable_collection
-from opendrop.utility.bindable.typing import ReadBindable
-from opendrop.utility.updaterworker import UpdaterWorker
-
-
-class FeatureExtractorParams:
-    def __init__(self) -> None:
-        self.bn_needle_region_px = VariableBindable(None)
-        self.bn_drop_region_px = VariableBindable(None)
-        self.bn_canny_min = VariableBindable(30)
-        self.bn_canny_max = VariableBindable(60)
-
-
-class FeatureExtractorService:
-    def extract_features(self, image: np.ndarray, params: FeatureExtractorParams):
-        return FeatureExtractor(image, params)
-
-
-class FeatureExtractor:
-    _Data = thread_safe_bindable_collection(
-        fields=[
-            'bn_edge_detection',
-            'bn_drop_profile_px',
-            'bn_needle_profile_px',
-            'bn_needle_width_px',
-        ]
-    )
-
-    def __init__(self, image: ReadBindable[np.ndarray], params: 'FeatureExtractorParams', *,
-                 loop: Optional[asyncio.AbstractEventLoop] = None) -> None:
-        self._loop = loop or asyncio.get_event_loop()
-
-        self._bn_image = image
-
-        self.params = params
-
-        self._data = self._Data(
-            _loop=self._loop,
-            bn_edge_detection=None,
-            bn_drop_profile_px=None,
-            bn_needle_profile_px=None,
-            bn_needle_width_px=math.nan,
-        )
-
-        self.is_busy = AccessorBindable(getter=self.get_is_busy)
-        self._updater_worker = UpdaterWorker(
-            do_update=self._update,
-            on_idle=self.is_busy.poke,
-            loop=self._loop,
-        )
-
-        self.bn_edge_detection = self._data.bn_edge_detection  # type: ReadBindable[Optional[np.ndarray]]
[…]
