### 2020-12-04 7f3f32a
Combine Physical Parameters with Image Processing pages in IFT
