### 2020-12-10 #2604 @jdber1
Hi @kyleniemeyer what is the status of this submission? I submitted to JOSS over 9 months ago, and would really like to get this out as soon as possible.
