### 2020-12-10 #2604 @kyleniemeyer
@jdber1 Sorry for the delay—I do understand your frustration, but please do understand that you submitted right at the beginning of a tough time. We do require two reviews, and due to the hardware requirement your submission is a bit more challenging. But, I have reached out again to the second reviewer.

Regarding figures and captions, [our docs](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) have some guidance on this, but I don't believe it is possible to control size/placement beyond placing the caption and figure commands:
```
Figures can be included like this:
![Caption for example figure.\label{fig:example}](figure.png)
and referenced from text using \autoref{fig:example}.
```

It does look like it's possible to vary the figure width, using (for example):
```
![Caption for example figure.\label{fig:example}](figure.png){ width=20% }
```
and the `\newpage` command does work.

But, otherwise, placement is up to Pandoc and LaTeX.
