+++ opendrop/fit/younglaplace/shape.py
@@ -1,0 +1,243 @@
+from functools import partial
+import typing
+from typing import Sequence, Tuple, Optional
+
+import numpy as np
+import scipy.integrate, scipy.optimize
+
+
+__all__ = ('YoungLaplaceShape',)
+
+INIT_SOLVED_SIZE = 4.0
+MAX_SOLVED_SIZE = 10.0
+
+MAX_CLOSEST_ITERATIONS = 10
+CLOSEST_TOL = 1.e-6
+
+# Very tiny float for managing singularities.
+INFITESIMAL = np.spacing(0.)
+
+
+class YoungLaplaceShape:
+    def __init__(self, bond: float) -> None:
+        self.bond = bond
+
+        self._s = [0.]
+        self._y = [(
+        #   r           z
+            0.        , 0.,
+        #   dr/ds       dz/ds
+            1.        , 0.,
+        #   dr/dBo      dz/dBo
+            0.        , 0.,
+        #   d2r/dBods  d2z/dBods
+            0.        , 0.,
+        )]
+        self._interpolants = []
+        self._z_inv_table = ([self._y[-1][1]], [self._s[-1]])
+        self._z_max_solved = False
+
+        self._solver = scipy.integrate.RK45(
+            fun=partial(young_laplace_ode_combined, bond=bond),
+            t0=self._s[-1],
+            y0=self._y[-1],
+            t_bound=MAX_SOLVED_SIZE,
+            # Using vectorized=False is faster for some reason.
+            vectorized=False,
+        )
+
+        self._step_until_s(INIT_SOLVED_SIZE)
+
+    def __call__(self, s: np.ndarray) -> np.ndarray:
+        r_and_z = self._eval(s)[[0,1]]
+        return r_and_z
+
+    def DBo(self, s: np.ndarray) -> np.ndarray:
+        drz_dBo = self._eval(s)[[4,5]]
+        return drz_dBo
+
+    def _eval(self, s: np.ndarray) -> np.ndarray:
+        result = np.empty(shape=(len(self._y[0]), len(s)))
+
+        s_abs = np.abs(s)
+
+        # Sort descending and then reverse so NaNs get accumulated at the start and the last element is always
+        # the largest.
+        order = np.argsort(-s_abs)[::-1]
+        unorder = np.empty_like(order)
+        unorder[order] = np.arange(len(order))
+
+        s_abs_ordered = s_abs[order]
+
+        self._step_until_s(s_abs[-1])
+
+        partition = np.searchsorted(s_abs_ordered, self._s[1:-1], side='right')
+        subintervals = np.split(s_abs_ordered, partition)
+        subresults = np.split(result, partition, axis=1)
+
+        for subresult, subinterval, func in zip(subresults, subintervals, self._interpolants):
+            subresult[:] = func(subinterval)
+
+        result = result[:, unorder]
+
+        # Flip signs of r, dz/ds, dr/dBo, d2z/dBods where negative s queried.
+        result[[0, 3, 4, 7], np.argwhere(s < 0)] *= -1
+
+        return result
+
+    def z_inv(self, z: Sequence[float]) -> np.ndarray:
+        result = np.interp(z, *self._z_inv_table, left=np.nan, right=np.nan) 
+
+        # Fast check for nan: https://stackoverflow.com/q/6736590.
+        if np.isnan(np.sum(result, where=(z > 0))) and not self._z_max_solved:
+            self._step_until_z(np.nanmax(z))
[…]
