+++ opendrop/fit/younglaplace/model.py
@@ -1,0 +1,443 @@
+# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
+# OpenDrop is released under the GNU GPL License. You are free to
+# modify and distribute the code, but always under the same license
+# (i.e. you cannot make commercial derivatives).
+#
+# If you use this software in your research, please cite the following
+# journal articles:
+#
+# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
+# R. F. Tabor, Measurement of surface and interfacial tension using
+# pendant drop tensiometry. Journal of Colloid and Interface Science 454
+# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
+#
+# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
+# and J. D. Berry, OpenDrop: Open-source software for pendant drop
+# tensiometry & contact angle measurements, submitted to the Journal of
+# Open Source Software
+#
+# These citations help us not only to understand who is using and
+# developing OpenDrop, and for what purpose, but also to justify
+# continued development of this code and other open source resources.
+#
+# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
+# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+# PURPOSE.  See the GNU General Public License for more details.  You
+# should have received a copy of the GNU General Public License along
+# with this software.  If not, see <https://www.gnu.org/licenses/>.
+
+
+import io
+import itertools
+import math
+import traceback
+from collections import namedtuple
+from enum import IntEnum
+from typing import Optional, Tuple, Iterable, overload, Any, Callable
+
+import numpy as np
+import scipy.optimize
+
+from .shape import YoungLaplaceShape
+
+
+MAXIMUM_FITTING_STEPS = 10
+
+SLOW_CONVERGENCE_THRESHOLD = 0.25
+FAST_CONVERGENCE_THRESHOLD = 0.75
+
+DELTA_TOL               = 1.e-6
+GRADIENT_TOL            = 1.e-6
+OBJECTIVE_TOL           = 1.e-4
+
+
+class _StopReason(IntEnum):
+    CONVERGENCE_IN_PARAMETERS = 1
+    CONVERGENCE_IN_GRADIENT = 2
+    CONVERGENCE_IN_OBJECTIVE = 4
+    MAXIMUM_STEPS_EXCEEDED = 8
+
+    @classmethod
+    def str_from_num(cls, v: int) -> str:
+        """Return a human-readable string representation of `v` (a flag field)."""
+        present_flag_names = []
+
+        for flag in cls:
+            if v & flag == flag:
+                present_flag_names.append(flag.name)
+
+        return ' | '.join(present_flag_names)
+
+
+# noinspection NonAsciiCharacters
+class YoungLaplaceFit:
+    _Params = namedtuple('Params', ('bond', 'radius', 'x', 'y', 'rotation'))
+
+    class _Cancelled(Exception):
+        pass
+
+    def __init__(
+            self,
+            data: np.ndarray,
+            *,
+            on_update: Optional[Callable[['YoungLaplaceFit'], Any]] = None,
+            logger: Optional[Callable[[str], Any]] = None
+    ) -> None:
+
+        self._data = data
+        self._params = self._Params(math.nan, math.nan, math.nan, math.nan, math.nan)
+
+        self._cancel_flag = False
+        self._is_cancelled = False
+        self._is_done = False
+
+        # The theoretical profile of the drop based on the current estimated parameters.
+        self._shape = None  # type: Optional[YoungLaplaceShape]
+        self._shape_size = 0.0
[…]
