### 2021-01-31 f5a3028
Allow C++ YoungLaplaceShape to accept np.ndarray as arguments
