### 2021-01-31 56954bd
Fix Cython scanner "pxd file from module name" finder
