### 2021-01-31 59ee8c1
Fix bug where YoungLaplaceShape.closest() would not work for points with z <= 0.
