### 2021-01-31 62396d9
Reimplement Young-Laplace fit parameter guess
The parameter guesser no longer depends on the order of the data points
passed in (the data points had to be in contiguous order along edge
previously).
