+++ opendrop/fit/younglaplace/model.py
@@ -51,6 +51,11 @@ GRADIENT_TOL            = 1.e-6
 OBJECTIVE_TOL           = 1.e-4
 
 
+# Math constants.
+PI = math.pi
+NAN = math.nan
+
+
 class _StopReason(IntEnum):
     CONVERGENCE_IN_PARAMETERS = 1
     CONVERGENCE_IN_GRADIENT = 2
@@ -85,7 +90,7 @@ class YoungLaplaceFit:
     ) -> None:
 
         self._data = data
-        self._params = self._Params(math.nan, math.nan, math.nan, math.nan, math.nan)
+        self._params = self._Params(NAN, NAN, NAN, NAN, NAN)
 
         self._cancel_flag = False
         self._is_cancelled = False
@@ -103,7 +108,8 @@ class YoungLaplaceFit:
 
     def _fit(self) -> None:
         # Guess the initial parameters
-        self._update_params(young_laplace_guess(self._data))
+        params = _young_laplace_guess(*self._data.T)
+        self._update_params(params)
 
         # Optimise
         try:
@@ -214,17 +220,20 @@ class YoungLaplaceFit:
         return λ_next, λ_cutoff_next, ssr_next, stop_reason
 
     def _calculate_jacobian(self) -> Tuple[np.ndarray, np.ndarray]:
-        O, w = ((self.apex_x,), (self.apex_y,)), self.rotation
+        O = [[self.apex_x],
+             [self.apex_y]]
+        w = self.rotation
         W = np.array([[np.cos(w), -np.sin(w)],
                       [np.sin(w),  np.cos(w)]])
 
         data_rz = W @ (self._data.T - O)
 
         s = self._shape.closest(data_rz/self.apex_radius)
+
         r, z = rz = self.apex_radius * self._shape(s)
         dr_dBo, dz_dBo = self.apex_radius * self._shape.DBo(s)
         e_r, e_z = e_rz = data_rz - rz
-        e = np.linalg.norm(e_rz, axis=0)
+        e = np.hypot(e_r, e_z)
 
         de_dBo = -(e_r*dr_dBo + e_z*dz_dBo) / e             # derivative w.r.t. Bo (Bond number)
         de_dR = -(e_r*r + e_z*z) / (self.apex_radius * e)   # derivative w.r.t. R (apex radius)
@@ -241,7 +250,7 @@ class YoungLaplaceFit:
 
         self._params = params
 
-        # Generate new drop shape when parameters change.
+        # Create a new drop shape when parameters change.
         self._shape = YoungLaplaceShape(params.bond)
 
         # Invoke update callback.
@@ -274,14 +283,14 @@ class YoungLaplaceFit:
     @property
     def volume(self) -> float:
         if self._shape is None:
-            return math.nan
+            return NAN
 
         return self._shape.volume(self._shape_size) * self.apex_radius**3
 
     @property
     def surface_area(self) -> float:
         if self._shape is None:
-            return math.nan
+            return NAN
 
         return self._shape.surface_area(self._shape_size) * self.apex_radius**2
 
@@ -297,9 +306,11 @@ class YoungLaplaceFit:
         except TypeError:
             return self.__call__([s])[0]
 
-        O, w = ((self.apex_x,), (self.apex_y,)), self.rotation
-        W = np.array([[ np.cos(w),  np.sin(w)],
-                      [-np.sin(w),  np.cos(w)]])
+        O = [[self.apex_x],
+             [self.apex_y]]
+        w = self.rotation
[…]
