+++ opendrop/fit/younglaplace/model.py
@@ -27,28 +27,13 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-import io
-import itertools
 import math
-import traceback
-from collections import namedtuple
-from enum import IntEnum
-from typing import Optional, Tuple, Iterable, overload, Any, Callable
+from typing import Optional, Sequence, Tuple
 
 import numpy as np
-import scipy.optimize
 
 from .shape import YoungLaplaceShape
-
-
-MAXIMUM_FITTING_STEPS = 10
-
-SLOW_CONVERGENCE_THRESHOLD = 0.25
-FAST_CONVERGENCE_THRESHOLD = 0.75
-
-DELTA_TOL               = 1.e-6
-GRADIENT_TOL            = 1.e-6
-OBJECTIVE_TOL           = 1.e-4
+from .types import YoungLaplaceParam
 
 
 # Math constants.
@@ -56,459 +41,131 @@ PI = math.pi
 NAN = math.nan
 
 
-class _StopReason(IntEnum):
-    CONVERGENCE_IN_PARAMETERS = 1
-    CONVERGENCE_IN_GRADIENT = 2
-    CONVERGENCE_IN_OBJECTIVE = 4
-    MAXIMUM_STEPS_EXCEEDED = 8
+class YoungLaplaceModel:
+    _shape: Optional[YoungLaplaceShape] = None
 
-    @classmethod
-    def str_from_num(cls, v: int) -> str:
-        """Return a human-readable string representation of `v` (a flag field)."""
-        present_flag_names = []
+    def __init__(self, data: Tuple[np.ndarray, np.ndarray]) -> None:
+        self.data = np.copy(data)
+        self.data.flags.writeable = False
 
-        for flag in cls:
-            if v & flag == flag:
-                present_flag_names.append(flag.name)
+        self._params = np.empty(len(YoungLaplaceParam))
+        self._s = np.empty(shape=(self.data.shape[1],))
+        self._residuals = np.empty(shape=(self.data.shape[1],))
+        self._jac = np.empty(shape=(self.data.shape[1], len(self._params)))
 
-        return ' | '.join(present_flag_names)
+    def set_params(self, params: Sequence[float]) -> None:
+        if (self._params == params).all():
+            return
 
+        bond   = params[YoungLaplaceParam.BOND]
+        radius = params[YoungLaplaceParam.RADIUS]
+        X0     = params[YoungLaplaceParam.APEX_X]
+        Y0     = params[YoungLaplaceParam.APEX_Y]
+        w      = params[YoungLaplaceParam.ROTATION]
 
-# noinspection NonAsciiCharacters
-class YoungLaplaceFit:
-    _Params = namedtuple('Params', ('bond', 'radius', 'x', 'y', 'rotation'))
+        s = self._s
+        residuals = self._residuals
+        de_dBo = self._jac[:, YoungLaplaceParam.BOND]
+        de_dR  = self._jac[:, YoungLaplaceParam.RADIUS]
+        de_dX0 = self._jac[:, YoungLaplaceParam.APEX_X]
+        de_dY0 = self._jac[:, YoungLaplaceParam.APEX_Y]
+        de_dw  = self._jac[:, YoungLaplaceParam.ROTATION]
 
-    class _Cancelled(Exception):
-        pass
-
-    def __init__(
-            self,
-            data: np.ndarray,
-            *,
-            on_update: Optional[Callable[['YoungLaplaceFit'], Any]] = None,
-            logger: Optional[Callable[[str], Any]] = None
-    ) -> None:
-
-        self._data = data
-        self._params = self._Params(NAN, NAN, NAN, NAN, NAN)
-
-        self._cancel_flag = False
-        self._is_cancelled = False
-        self._is_done = False
-
[…]
