+++ opendrop/processing/ift/needle_width.py
@@ -1,138 +1,0 @@
-# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
-# OpenDrop is released under the GNU GPL License. You are free to
-# modify and distribute the code, but always under the same license
-# (i.e. you cannot make commercial derivatives).
-#
-# If you use this software in your research, please cite the following
-# journal articles:
-#
-# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
-# R. F. Tabor, Measurement of surface and interfacial tension using
-# pendant drop tensiometry. Journal of Colloid and Interface Science 454
-# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
-#
-# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
-# and J. D. Berry, OpenDrop: Open-source software for pendant drop
-# tensiometry & contact angle measurements, submitted to the Journal of
-# Open Source Software
-#
-# These citations help us not only to understand who is using and
-# developing OpenDrop, and for what purpose, but also to justify
-# continued development of this code and other open source resources.
-#
-# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
-# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-# PURPOSE.  See the GNU General Public License for more details.  You
-# should have received a copy of the GNU General Public License along
-# with this software.  If not, see <https://www.gnu.org/licenses/>.
-
-
-import itertools
-import math
-from typing import Tuple
-
-import numpy as np
-
-NEEDLE_TOL = 1.e-4
-NEEDLE_STEPS = 20
-
-
-def calculate_width_from_needle_profile(needle_profile: Tuple[np.ndarray, np.ndarray]) -> float:
-    """Return the width of the needle defined by `needle_profile`
-
-    :param needle_contours: The two edges of the needle, this is an `np.ndarray` in the format:
-                                [[[x0, y0], ..., [xn, yn]], [[x'0, y'0], ..., [x'm, y'm]]]
-                            Where xi, yi correspond to the left edge of the needle and x'i, y'i, the right edge.
-    :return: Diameter of the needle (in same units as `needle_profile`)
-    """
-    if len(needle_profile) != 2:
-        raise ValueError(
-            "'needle_profile' must contain only two edges (left and right edge of needle), got '{}' edge(s)"
-            .format(len(needle_profile))
-        )
-
-    if len(needle_profile[0]) == 0 or len(needle_profile[1]) == 0:
-        return math.nan
-
-    # Sort the points of each contour in ascending y coordinate.
-    needle_profile = tuple(x[x[:, 1].argsort()] for x in needle_profile)  # type: Tuple[np.ndarray, np.ndarray]
-
-    p0 = needle_profile[0][0]
-
-    # Translate the needle contours such that the "top-left" point is at (0, 0)
-    needle_profile = tuple(x - p0 for x in needle_profile)
-
-    x0, x1, theta = _optimise_needle(needle_profile)
-
-    needle_diameter = abs((x1 - x0) * np.sin(theta))
-
-    return needle_diameter
-
-
-def _optimise_needle(needle_profile: Tuple[np.ndarray, np.ndarray]) -> Tuple[float, float, float]:
-    """
-    :param needle_profile: The two edges of the needle, see `calculate_needle_width_from_profile()` doc
[…]
