### 2021-02-04 2dde8f3
Delete PendantAnalysisJob waiting for image after cancelling
