### 2021-02-04 589724c
Reimplement _general_purpose_equality_check
Old implementation would throw numpy deprecation errors about creating
ragged arrays.
