### 2021-02-04 63e9ab4
Refactor USBCamera acquisition
