+++ opendrop/app/common/image_acquisition/configurator/usb_camera/__init__.py
@@ -27,4 +27,4 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
-from .component import usb_camera_cs
+from . import usb_camera, chooser
