+++ opendrop/app/common/image_acquisition/configurator/usb_camera/component.py
@@ -1,363 +1,0 @@
-# Copyright © 2020, Joseph Berry, Rico Tabor (opendrop.dev@gmail.com)
-# OpenDrop is released under the GNU GPL License. You are free to
-# modify and distribute the code, but always under the same license
-# (i.e. you cannot make commercial derivatives).
-#
-# If you use this software in your research, please cite the following
-# journal articles:
-#
-# J. D. Berry, M. J. Neeson, R. R. Dagastine, D. Y. C. Chan and
-# R. F. Tabor, Measurement of surface and interfacial tension using
-# pendant drop tensiometry. Journal of Colloid and Interface Science 454
-# (2015) 226–237. https://doi.org/10.1016/j.jcis.2015.05.012
-#
-# E. Huang, T. Denning, A. Skoufis, J. Qi, R. R. Dagastine, R. F. Tabor
-# and J. D. Berry, OpenDrop: Open-source software for pendant drop
-# tensiometry & contact angle measurements, submitted to the Journal of
-# Open Source Software
-#
-# These citations help us not only to understand who is using and
-# developing OpenDrop, and for what purpose, but also to justify
-# continued development of this code and other open source resources.
-#
-# OpenDrop is distributed WITHOUT ANY WARRANTY; without even the
-# implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
-# PURPOSE.  See the GNU General Public License for more details.  You
-# should have received a copy of the GNU General Public License along
-# with this software.  If not, see <https://www.gnu.org/licenses/>.
-
-
-from typing import Optional, Callable, Any
-
-from gi.repository import Gtk, Gdk, GObject
-
-from opendrop.app.common.services.acquisition import USBCameraAcquirer
-from opendrop.mvp import ComponentSymbol, Presenter, View
-from opendrop.utility.bindable.typing import Bindable
-from opendrop.utility.bindable.gextension import GObjectPropertyBindable
-from opendrop.widgets.float_entry import FloatEntry
-from opendrop.widgets.integer_entry import IntegerEntry
-
-usb_camera_cs = ComponentSymbol()  # type: ComponentSymbol[Gtk.Widget]
-
-
-@usb_camera_cs.view()
-class USBCameraView(View['USBCameraPresenter', Gtk.Widget]):
-    STYLE = '''
-    .change-cam-dialog-view-footer {
-         background-color: gainsboro;
-    }
-
-    .small-pad {
-         min-height: 0px;
-         min-width: 0px;
-         padding: 6px 4px 6px 4px;
-    }
-
-    .dialog-footer-button {
-         min-height: 0px;
-         min-width: 0px;
-         padding: 8px 6px 8px 6px;
-    }
-
-    .error {
-        color: red;
-        border: 1px solid red;
-    }
-
-    .error-text {
-        color: red;
-    }
-    '''
-
-    _STYLE_PROV = Gtk.CssProvider()
-    _STYLE_PROV.load_from_data(bytes(STYLE, 'utf-8'))
-    Gtk.StyleContext.add_provider_for_screen(Gdk.Screen.get_default(), _STYLE_PROV, Gtk.STYLE_PROVIDER_PRIORITY_USER)
-
-    def _do_init(self) -> Gtk.Widget:
-        self._change_camera_dialog_cid = None  # type: Any
-
-        self._widget = Gtk.Grid(row_spacing=10, column_spacing=10)
-
-        # Populate self.widget
-
-        camera_container = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=5)
-        self._widget.attach(camera_container, 0, 0, 2, 1)
-
[…]
