### 2021-02-04 fa565be
Make genicam an optional dependency
