### 2021-02-04 27e8415
Fix typing_extensions ModuleNotFoundError on Python 3.9
