### 2021-02-05 e0a8aa3
Fix Python 3.6 compatibility
functools.lru_cache changed api in Python 3.8
