### 2021-02-05 0100da0
Update docs to drop Python 3.5 support
