+++ README.rst
@@ -7,7 +7,7 @@
 
 **OpenDrop** is a fully-featured pendant drop tensiometry software, allowing acquisition, analysis and fitting of pendant drop profiles to obtain surface and interfacial tension. OpenDrop also includes functionality for measuring sessile drop contact angles.
 
-It runs on **Python 3.5+**, and uses **GTK+ 3** for the user interface, and **OpenCV** for image analysis.
+It runs on **Python 3.6+**, and uses **GTK+ 3** for the user interface, and **OpenCV** for image analysis.
 
 OpenDrop is released under the **GNU GPL** open source license, and available for free.
 
