### 2021-02-06 2d1fbf6
Move opendrop.utility.geometry to opendrop.geometry
