+++ opendrop/features/colorize.pyx
@@ -1,0 +1,52 @@
+from libc.string cimport memcpy
+cimport cython
+cimport numpy as np
+from cpython cimport array
+import array
+
+
+__all__ = ('colorize_labels',)
+
+
+ctypedef fused integral:
+    np.int8_t
+    np.uint8_t
+    np.int16_t
+    np.uint16_t
+    np.int32_t
+    np.uint32_t
+    np.int64_t
+    np.uint64_t
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+def colorize_labels(integral[:, :] labels not None, np.uint8_t[:, ::1] colors not None):
+    cdef size_t i, j, n
+    cdef array.array arr
+    cdef np.uint8_t[:, :, :] view
+    
+    if colors.shape[1] != 4:
+        raise ValueError(
+            "axis 1 of colors array must equal 4, got {}"
+            .format(colors.shape[1])
+        )
+    
+    arr = array.array('B')
+    array.resize(arr, labels.shape[0]*labels.shape[1]*4)
+    
+    view = <np.uint8_t[:labels.shape[0], :labels.shape[1], :4]> arr.data.as_uchars
+
+    for i in range(labels.shape[0]):
+        for j in range(labels.shape[1]):
+            n = labels[i][j]
+            
+            if n >= colors.shape[0]:
+                raise IndexError(
+                    "index {} is out of bounds for axis 0 of colors array with shape {}"
+                    .format(n, (colors.shape[0], colors.shape[1]))
+                )
+            
+            memcpy(<void *>&view[i][j][0], <void *>&colors[n][0], 4)
+
+    return arr
