+++ opendrop/features/__init__.py
@@ -1,0 +1,2 @@
+from .colorize import *
+from .pendant import *
