### 2021-02-07 a9814dc
Fix bug where all AffineFrameArtists shared a transform matrix
