### 2021-02-07 e23bf28
Add missing check for when _invalidate_handler is None in invalidate()
