### 2021-02-07 545a52f
Add correct handling for when region is None
