### 2021-02-07 b1dfad9
Combine invalidate_region() and invalidate()
