### 2021-02-07 1e64397
Fix potential race-ish condition
When zoom() is called, the canvas transformation matrix is not
immediately recalculated, can lead to flickering of FLOATING attached
artists when canvas is being invalidated/updated frequently.
