### 2021-02-08 275f8dc
Fix bug where selected images can't be changed after image processing
