### 2021-02-08 b829801
Sort child artists by z-index before drawing in AffineFrameArtist
