### 2021-02-09 99f9e2a
Guarantee AffineFrameArtist draws last added child on top
