### 2021-02-09 1e8af39
Draw focus border when Canvas is focused
