### 2021-02-09 7c61f8a
Remove adding default event masks to Canvas
