### 2021-02-09 873cf5b
Add a bunch of canvas Artists
