+++ opendrop/widgets/canvas/__init__.py
@@ -1,3 +1,9 @@
 from ._canvas import *
 from ._artist import *
 from ._affine import *
+from ._rectangle import *
+from ._circle import *
+from ._image import *
+from ._polyline import *
+from ._line import *
+from ._angle import *

[…]
