### 2021-02-09 cc87bfd
Retire Render and switch to new Canvas widget
