+++ opendrop/app/conan/report/overview/detail/image.py
@@ -32,11 +32,10 @@ from typing import Optional
 
 from gi.repository import Gtk, GObject
 
-from opendrop.utility.gmisc import pixbuf_from_array
-from opendrop.geometry import Rect2
-from opendrop.widgets.render import Render
-from opendrop.widgets.render.objects import PixbufFill, Line, Polyline, Angle
 from opendrop.appfw import Presenter, component, install
+from opendrop.geometry import Rect2, Vector2
+from opendrop.widgets.canvas import Canvas, CanvasAlign, ImageArtist, LineArtist, PolylineArtist, AngleLabelArtist
+
 from opendrop.app.conan.analysis import ConanAnalysis
 
 
@@ -44,38 +43,64 @@ from opendrop.app.conan.analysis import ConanAnalysis
     template_path='./image.ui',
 )
 class ConanReportOverviewImagePresenter(Presenter[Gtk.Container]):
-    _analysis = None  # type: Optional[ConanAnalysis]
+    _analysis: Optional[ConanAnalysis] = None
     event_connections = ()
 
     def after_view_init(self) -> None:
-        self.render = Render(visible=True)
-        self.host.add(self.render)
+        self.canvas = Canvas(align=CanvasAlign.FIT, hexpand=True, vexpand=True, visible=True)
+        self.host.add(self.canvas)
 
-        self.background_ro = PixbufFill()
-        self.render.add_render_object(self.background_ro)
+        self.bg_artist = ImageArtist()
+        self.canvas.add_artist(self.bg_artist)
 
-        self.surface_line_ro = Line(stroke_color=(0.25, 1.0, 0.25))
-        self.render.add_render_object(self.surface_line_ro)
+        self.baseline_artist = LineArtist(
+            stroke_color=(0.25, 1.0, 0.25),
+            scale_strokes=True,
+        )
+        self.canvas.add_artist(self.baseline_artist)
 
-        self.drop_contour_ro = Polyline(stroke_color=(0.0, 0.5, 1.0))
-        self.render.add_render_object(self.drop_contour_ro)
+        self.drop_edge_artist = PolylineArtist(
+            stroke_color=(0.0, 0.5, 1.0),
+            scale_strokes=True,
+        )
+        self.canvas.add_artist(self.drop_edge_artist)
 
-        self.left_angle_ro = Angle(stroke_color=(1.0, 0.0, 0.5), clockwise=False)
-        self.render.add_render_object(self.left_angle_ro)
+        self.left_angle_artist = AngleLabelArtist(
+            clockwise=False,
+            arc_radius=20.0,
+            text_radius=40.0,
+            end_line_radius=50.0,
+            stroke_color=(1.0, 0.0, 0.5),
+            text_color=(1.0, 0.0, 0.5),
+            scale_radii=True,
+            scale_strokes=True,
+            scale_text=True,
+        )
+        self.canvas.add_artist(self.left_angle_artist)
 
-        self.right_angle_ro = Angle(stroke_color=(1.0, 0.0, 0.5), clockwise=True)
-        self.render.add_render_object(self.right_angle_ro)
+        self.right_angle_artist = AngleLabelArtist(
+            clockwise=True,
+            arc_radius=20.0,
+            text_radius=40.0,
+            end_line_radius=50.0,
+            stroke_color=(1.0, 0.0, 0.5),
+            text_color=(1.0, 0.0, 0.5),
+            scale_radii=True,
+            scale_strokes=True,
+            scale_text=True,
+        )
+        self.canvas.add_artist(self.right_angle_artist)
 
-        self.surface_line_ro.bind_property(
+        self.baseline_artist.bind_property(
             'line',
-            self.left_angle_ro, 'start-angle',
+            self.left_angle_artist, 'start-angle',
             GObject.BindingFlags.SYNC_CREATE,
[…]
