### 2021-02-09 dfe8327
Remove unused opendrop.utility.worker module
