### 2021-02-09 fb00b70
Remove unused pixbuf_from_array() function
