### 2021-02-09 62d28f8
Remove unused cairo_saved() context manager
