### 2021-02-09 45098ca
Add scale-radius property to CircleArtist
