+++ opendrop/widgets/canvas/_circle.py
@@ -17,27 +17,37 @@ class CircleArtist(Artist):
     _xc: float = 0.0
     _yc: float = 0.0
     _radius: float = 0.0
+    _scale_radius = False
     _fill_color = (0.0, 0.0, 0.0)
 
     _last_drawn_region: Optional[cairo.Region] = None
+    _last_drawn_radius_scale = 1.0
 
     def draw(self, cr: cairo.Context) -> None:
         xc = self._xc
         yc = self._yc
         radius = self._radius
+        scale_radius = self._scale_radius
         fill_color = self._fill_color
 
         if radius == 0.0:
             return
 
-        dx = 1/cr.get_matrix().xx
-        dy = 1/cr.get_matrix().yy
+        matrix = cr.get_matrix()
+        dx = 1/matrix.xx
+        dy = 1/matrix.yy
 
-        cr.arc(xc, yc, radius, 0, 2*math.pi)
+        if scale_radius:
+            radius_scale = 1/(0.5 * (matrix.xx + matrix.yy))
+        else:
+            radius_scale = 1.0
+
+        cr.arc(xc, yc, radius_scale*radius, 0, 2*math.pi)
         cr.set_source_rgba(*fill_color)
         cr.fill()
 
-        self._last_drawn_region = circle_region(xc, yc, radius + max(dx, dy))
+        self._last_drawn_region = circle_region(xc, yc, radius_scale*radius + max(dx, dy))
+        self._last_drawn_radius_scale = radius_scale
 
     @GObject.Property
     def extents(self) -> Optional[Rect2[float]]:
@@ -48,7 +58,6 @@ class CircleArtist(Artist):
         self._extents = extents
         self._invalidate()
 
-    # Style properties
     @GObject.Property(type=float, default=_xc)
     def xc(self) -> float:
         return self._xc
@@ -76,6 +85,15 @@ class CircleArtist(Artist):
         self._radius = math.fabs(radius)
         self._invalidate()
 
+    @GObject.Property(type=bool, default=_scale_radius)
+    def scale_radius(self) -> bool:
+        return self._scale_radius
+
+    @scale_radius.setter
+    def scale_radius(self, value: bool) -> None:
+        self._scale_radius = value
+        self._invalidate()
+
     @GObject.Property
     def fill_color(self) -> Sequence[float]:
         return self._fill_color
@@ -89,8 +107,9 @@ class CircleArtist(Artist):
         xc = self._xc
         yc = self._yc
         radius = self._radius
+        radius_scale = self._last_drawn_radius_scale
 
         inv_region = self._last_drawn_region or cairo.Region()
-        inv_region.union(circle_region(xc, yc, radius))
+        inv_region.union(circle_region(xc, yc, radius_scale*radius))
 
         self.invalidate(inv_region)
