### 2021-02-09 3e10935
Scale UI elements of line definer
