### 2021-02-09 757e75e
Move UI specific initialization tasks to opendrop.app.__init__
