### 2021-02-09 9a1294b
Scale PolylineArtist strokes in preview
