### 2021-02-09 1a51508
Edit IFT graph labels in UI for consistency
