### 2021-02-09 jdber1/opendrop#28 @FelixKratz
Saving Data to Disk not working
I seem to have an issue with saving the data I get from the analysis to disk.
I have installed the newest version of the application alongside with python@3.9 on macOS.

The error is:
  File "/usr/local/lib/python3.9/site-packages/opendrop/app/ift/analysis_saver/component.py", line 251, in _get_save_dir_parent
    path_str = self._save_dir_parent_inp.get_filename()
SystemError: gi.FunctionInfo(get_filename) returned a result with an error set

EDIT: Noticed that this is probably related:
When I type the output name into the corresponding textbox the following error arises with every keystroke:
  File "/usr/local/lib/python3.9/site-packages/opendrop/utility/bindable/misc.py", line 8, in _general_purpose_equality_check
    if isinstance(x, numbers.Number) and isinstance(y, numbers.Number):
RecursionError: maximum recursion depth exceeded in __instancecheck__
