### 2021-02-09 jdber1/opendrop#30 @FelixKratz
Application will stop analysing images
When testing with ~200 images the application does not seem to analyse the images and an ever spinning progress indicator is shown. When selecting only ~50 images the analysis works as expected.  Could this have to do with multiprocessing? I experience this issue under macOS and Linux on a 6 core CPU.
