### 2021-02-09 f79e9c4
Reduce priority of spinner animation timeout sources
Too many spinner timeouts at DEFAULT priority prevents callbacks
scheduled with call_soon() to be invoked, (like fitting functions
offloaded to separate processes).
