### 2021-02-09 jdber1/opendrop#30 @eugenhu
It seems like when there were too many spinners, the event loop was just repeatedly invoking the spinners' animation update cycles instead of the callbacks to tell the program that fitting had completed since the animation callbacks had the same 'default' priority.

Hopefully this is fixed with https://github.com/jdber1/opendrop/commit/f79e9c4c8aaf68f7df0cd52c5fa424b8ae8db452.

Update with
```
pip install -U git+https://github.com/jdber1/opendrop.git
```
Thank you for the prompt feedback, I'll work on addressing two of your other issues tomorrow.
