### 2021-02-10 402e54f
Reimplement _general_purpose_equality_check
Currently defaults to using 'is' to check for equality when x and y are
not numbers/arrays. This does not work for non-interned strings and
other types without unique instances. This can cause an infinite loop in
Bindings.
