### 2021-02-10 #2604 @FelixKratz
We have gone through the paper and tested the application. The problems we pointed out in [Application will stop analysing images](https://github.com/jdber1/opendrop/issues/30) and [Saving Data to Disk not working](https://github.com/jdber1/opendrop/issues/28) where resolved very quickly. We have compared the results of the presented application with two other methods and find consistent predictions. We thus recommend the paper and the application.
