### 2021-02-10 #2604 @kyleniemeyer
OK @jdber1, with two reviewers now recommending acceptance, we can now proceed to publication! I will do some final checks on your paper, but in the meantime please archive the software repo (e.g., at Zenodo) and provide the DOI here. And, thanks for your patience with this review.
