### 2021-02-11 c1720a5
Fix needle_guess() bug
Distances of needle pixels from center was incorrectly being calculated
(wrong axis used for np.linalg.norm)
