### 2021-02-11 ac371ad
Add more robust needle fitting algorithm
needle_guess() can guess correct parameters even with outliers, but it's
also a lot slower.
