+++ opendrop/fit/needle/model.py
@@ -12,39 +12,39 @@ class NeedleModel:
         self._params = np.empty(len(NeedleParam))
         self._residuals = np.empty(shape=(self.data.shape[1],))
         self._jac = np.empty(shape=(self.data.shape[1], len(self._params)))
+        self._lmask = np.empty(shape=(self.data.shape[1],), dtype=bool)
 
     def set_params(self, params: Sequence[float]) -> None:
         if (self._params == params).all():
             return
 
         w      = params[NeedleParam.ROTATION]
+        rho    = params[NeedleParam.RHO]
         radius = params[NeedleParam.RADIUS]
-        X0     = params[NeedleParam.CENTER_X]
-        Y0     = params[NeedleParam.CENTER_Y]
 
         residuals = self._residuals
-        de_dw  = self._jac[:, NeedleParam.ROTATION]
-        de_dR  = self._jac[:, NeedleParam.RADIUS]
-        de_dX0 = self._jac[:, NeedleParam.CENTER_X]
-        de_dY0 = self._jac[:, NeedleParam.CENTER_Y]
+        de_dw   = self._jac[:, NeedleParam.ROTATION]
+        de_drho = self._jac[:, NeedleParam.RHO]
+        de_dR   = self._jac[:, NeedleParam.RADIUS]
+        lmask   = self._lmask
 
         Q = np.array([[np.cos(w), -np.sin(w)],
                       [np.sin(w),  np.cos(w)]])
 
         data_x, data_y = self.data
-        data_r, data_z = Q.T @ (data_x - X0, data_y - Y0)
+        data_r, data_z = Q.T @ (data_x, data_y) - [[rho], [0]]
 
         e = np.abs(data_r) - radius
 
-        rmask = data_r >= 0
-        lmask = ~rmask
+        lmask[:] = data_r < 0
+        rmask    = ~lmask
 
         residuals[:] = e
         de_dw[rmask] =  data_z[rmask]
         de_dw[lmask] = -data_z[lmask]
         de_dR[:] = -1
-        de_dX0[rmask], de_dY0[rmask] = -Q.T[0]
-        de_dX0[lmask], de_dY0[lmask] =  Q.T[0]
+        de_drho[rmask] = -1
+        de_drho[lmask] =  1
 
         self._params[:] = params
 
@@ -69,3 +69,9 @@ class NeedleModel:
         residuals = self._residuals[:]
         residuals.flags.writeable = False
         return residuals
+
+    @property
+    def lmask(self) -> np.ndarray:
+        lmask = self._lmask[:]
+        lmask.flags.writeable = False
+        return lmask
