+++ opendrop/fit/needle/types.py
@@ -2,6 +2,5 @@ from enum import IntEnum, auto
 
 class NeedleParam(IntEnum):
     ROTATION = 0
+    RHO      = auto()
     RADIUS   = auto()
-    CENTER_X = auto()
-    CENTER_Y = auto()
