+++ opendrop/fit/needle/guess.py
@@ -1,41 +1,57 @@
 from typing import Sequence
+import math
 
 import numpy as np
+import scipy.signal
+
+from opendrop.geometry import Rect2
 
 from .types import NeedleParam
+from .hough import hough
+
+
+_rng = np.random.default_rng()
 
 
 def needle_guess(data: np.ndarray) -> Sequence[float]:
     params = np.empty(len(NeedleParam))
+    data = data.astype(float)
 
-    cm = data.mean(axis=1)
-    # Points relative to "centre of mass".
-    data_cm = data - cm.reshape(2, 1)
-    dists = np.linalg.norm(data_cm, axis=0)
-    closest_cm = data_cm[:, np.argsort(dists)[0]]
-    projection = (closest_cm.reshape(2, 1) * data_cm).sum(axis=0)
+    if data.shape[1] > 200:
+        data = _rng.choice(data, size=100, replace=False, axis=1)
 
-    edge1_mask = projection > 0
-    edge1 = data[:, edge1_mask]
-    edge1_dir = _guess_line_direction(edge1)
-    edge1_perp = edge1_dir[::-1] * [1, -1]
+    extents = Rect2(data.min(axis=1), data.max(axis=1))
+    diagonal = int(math.ceil((extents.w**2 + extents.h**2)**0.5))
+    data -= np.reshape(extents.center, (2, 1))
+    votes = hough(data, diagonal)
 
-    rotation = -np.arctan2(*edge1_dir)
-    radius = np.abs(edge1_perp @ closest_cm.reshape(2, 1))
-    center_x = cm[0]
-    center_y = cm[1]
+    needles = np.zeros(shape=(votes.shape[0], 3))
+    for i in range(votes.shape[0]):
+        peaks, props = scipy.signal.find_peaks(votes[i], prominence=0)
+        if len(peaks) < 2: continue
+        ix = np.argsort(props['prominences'])[::-1]
+        peak1_i, peak2_i = peaks[ix[:2]]
+        prom1, prom2 = props['prominences'][ix[:2]]
+        if prom2 < prom1/2: continue
 
-    params[NeedleParam.ROTATION] = rotation
+        peak1 = ((peak1_i - 1)/(len(votes[i]) - 3) - 0.5) * diagonal
+        peak2 = ((peak2_i - 1)/(len(votes[i]) - 3) - 0.5) * diagonal
+
+        needles[i][0] = (peak1 + peak2)/2
+        needles[i][1] = math.fabs(peak1 - peak2)/2
+        needles[i][2] = prom1 + prom2
+
+    scores = scipy.ndimage.gaussian_filter(needles[:, 2], sigma=10, mode='wrap')
+    needle_i = scores.argmax()
+    
+    theta = (needle_i/len(needles)) * np.pi + np.pi/2
+    rho, radius = needles[needle_i][:2]
+
+    rho_offset = np.cos(theta)*extents.xc - np.sin(theta)*extents.yc
+    rho += rho_offset
+
+    params[NeedleParam.ROTATION] = theta
+    params[NeedleParam.RHO] = rho
     params[NeedleParam.RADIUS] = radius
-    params[NeedleParam.CENTER_X] = center_x
-    params[NeedleParam.CENTER_Y] = center_y
 
     return params
-
-
-def _guess_line_direction(data: np.ndarray) -> np.ndarray:
-    x0 = data[:, 0]
-    dists = np.argsort(np.linalg.norm(data - x0.reshape(2, 1), axis=0))
-    x1 = data[:, dists[-1]]
-
-    return (x1 - x0)/np.linalg.norm(x1 - x0)
