### 2021-02-11 b12a2d7
Rename detect_pendant_features() to label_pendant_edges()
