### 2021-02-11 b862aa7
Refactor opendrop.fit module
