+++ opendrop/app/ift/services/younglaplace.py
@@ -4,7 +4,7 @@ from typing import Tuple
 
 import numpy as np
 
-from opendrop.fit.younglaplace import YoungLaplaceFitResult, young_laplace_fit
+from opendrop.fit import YoungLaplaceFitResult, young_laplace_fit
 
 
 __all__ = ('YoungLaplaceFitResult', 'YoungLaplaceFitService')
