+++ opendrop/fit/younglaplace/guess.py
@@ -2,8 +2,8 @@ import math
 from typing import Tuple
 
 import numpy as np
-import scipy.optimize
 
+from ..pendant import find_pendant_apex
 from .types import YoungLaplaceParam
 
 
@@ -15,111 +15,21 @@ NAN = math.nan
 def young_laplace_guess(data: Tuple[np.ndarray, np.ndarray]) -> tuple:
     params = np.empty(len(YoungLaplaceParam))
 
-    r, z, radius, apex_x, apex_y, rotation = _guess_rz_from_xy(*data)
+    apex, radius, rotation = find_pendant_apex(data)
+
+    r, z = np.array([[np.cos(rotation), -np.sin(rotation)],
+                     [np.sin(rotation),  np.cos(rotation)]]) \
+           @ data
+
     bond = _bond_selected_plane(r, z, radius)
 
     params[YoungLaplaceParam.BOND] = bond
     params[YoungLaplaceParam.RADIUS] = radius
-    params[YoungLaplaceParam.APEX_X] = apex_x
-    params[YoungLaplaceParam.APEX_Y] = apex_y
+    params[YoungLaplaceParam.APEX_X] = apex.x
+    params[YoungLaplaceParam.APEX_Y] = apex.y
     params[YoungLaplaceParam.ROTATION] = rotation
     
     return params
-
-
-def _guess_rz_from_xy(x: np.ndarray, y: np.ndarray) -> tuple:
-    xc = x.mean()
-    yc = y.mean()
-    radius = np.hypot(x - xc, y - yc).mean()
-
-    # Fit a circle to the most circular part of the data.
-    ans = scipy.optimize.least_squares(
-        _circle_residues,
-        (xc, yc, radius),
-        _circle_jac,
-        args=(x, y),
-        loss='cauchy',  # Ignore outliers.
-        f_scale=radius/100,
-        x_scale=(radius, radius, radius),
-    )
-    xc, yc, radius = ans.x
-    resids = np.abs(ans.fun)
-    resids_50ptile = np.quantile(resids, 0.5)
-
-    # The somewhat circular-ish part of the drop profile.
-    bowl_mask = resids < 10*resids_50ptile
-    bowl_x = x[bowl_mask]
-    bowl_y = y[bowl_mask]
-
-    # Don't need these variables anymore.
-    del resids, resids_50ptile
-
-    # Find the symmetry axis of bowl.
-    Ixx, Iyy, Ixy = _calculate_inertia(bowl_x, bowl_y)
-
-    # Eigenvector calculation for a symmetric 2x2 matrix.
-    rotation = 0.5 * np.arctan2(2 * Ixy, Ixx - Iyy)
-    unit_r = np.array([ np.cos(rotation), np.sin(rotation)])
-    unit_z = np.array([-np.sin(rotation), np.cos(rotation)])
-
-    bowl_r = unit_r @ [bowl_x - xc, bowl_y - yc]
-    bowl_z = unit_z @ [bowl_x - xc, bowl_y - yc]
-    bowl_r_ix = np.argsort(bowl_r)
-    bowl_z_ix = np.argsort(bowl_z)
-
-    # Calculate "asymmetry" along each axis. We define this to be the squared difference between the left and
-    # right points, integrated along the axis.
-    ma_kernel = np.ones(len(bowl_r)//10)/(len(bowl_r)//10)
-    asymm_r = (np.convolve((bowl_r - bowl_r.mean())[bowl_r_ix], ma_kernel, mode='valid')**2).sum()
-    asymm_z = (np.convolve((bowl_z - bowl_z.mean())[bowl_z_ix], ma_kernel, mode='valid')**2).sum()
-    if asymm_z > asymm_r:
[…]
