### 2021-02-12 #2604 @jdber1
> I will do some final checks on your paper, but in the meantime please archive the software repo (e.g., at Zenodo) and provide the DOI here

Thanks very much @kyleniemeyer! I will archive the repo and report the DOI when I get a chance.
