### 2021-02-12 15aeeaf
Add rotation_mat2d() function
