+++ opendrop/utility/__init__.py
@@ -27,3 +27,4 @@
 # with this software.  If not, see <https://www.gnu.org/licenses/>.
 
 
+from .misc import *
