### 2021-02-12 0bd9839
Wordsmith README and fix docs index format
