+++ README.rst
@@ -5,28 +5,23 @@
 
 |
 
-**OpenDrop** is a fully-featured pendant drop tensiometry software, allowing acquisition, analysis and fitting of pendant drop profiles to obtain surface and interfacial tension. OpenDrop also includes functionality for measuring sessile drop contact angles.
+**OpenDrop** is a fully-featured image analysis software for performing pendant drop tensiometry and contact angle measurements. Images can be loaded from the file system or acquired directly from USB webcams or GenICam (GigE Vision, USB3 Vision) compliant industrial cameras.
 
-It runs on **Python 3.6+**, and uses **GTK+ 3** for the user interface, and **OpenCV** for image analysis.
-
-OpenDrop is released under the **GNU GPL** open source license, and available for free.
+The software is released under the **GNU GPL** open source license, and available for free.
 
 *********
 Installation & User Guide
 *********
-**OpenDrop** currently runs on Ubuntu, MacOS and Windows 10 operating systems.
+OpenDrop is currently distributed as a Python package and can be installed on most operating systems (Windows, macOS, Linux).
+
 For installation instructions and user guide, visit: https://opendrop.readthedocs.io/
 
-*********
-Examples
-*********
-
-Example images are located in the example_images subfolder. Instructions on how to process the pendant drop image to extract surface tension, and the images of drops on surfaces to extract contact angle, are also available at: https://opendrop.readthedocs.io/
+Example images have been provided in the 'example_images' folder.
 
 *********
 Support
 *********
-For any questions, issues, problems, or feedback feel free to open an issue here: https://github.com/jdber1/opendrop/issues
+For any questions, issues, or feedback feel free to `open an issue on the GitHub repo <https://github.com/jdber1/opendrop/issues>`_.
 
-We can also be contacted by email `here <mailto:opendrop.dev@gmail.com>`_
+We can also be contacted by email `here <mailto:opendrop.dev@gmail.com>`_.
 
