### 2021-02-12 0abaa4b
Update Getting Started doc
