### 2021-02-12 2fc06f9
Add a screenshot to README
