+++ README.rst
@@ -7,6 +7,10 @@
 
 **OpenDrop** is a fully-featured image analysis software for performing pendant drop tensiometry and contact angle measurements. Images can be loaded from the file system or acquired directly from USB webcams or GenICam (GigE Vision, USB3 Vision) compliant industrial cameras.
 
+.. raw:: html
+
+    <img src="https://opendrop.readthedocs.io/en/latest/_images/ift_image_processing_regions_defined.png" width="60%">
+
 The software is released under the **GNU GPL** open source license, and available for free.
 
 *********
@@ -14,7 +18,7 @@ Installation & User Guide
 *********
 OpenDrop is currently distributed as a Python package and can be installed on most operating systems (Windows, macOS, Linux).
 
-For installation instructions and user guide, visit: https://opendrop.readthedocs.io/
+For installation instructions and user guides, visit: https://opendrop.readthedocs.io/
 
 Example images have been provided in the 'example_images' folder.
 
