### 2021-02-22 55afd2b
Fix _general_purpose_equality_check
Make it recursiveley equate elements of iterables.
