### 2021-02-22 #2604 @jdber1
> OK @jdber1, with two reviewers now recommending acceptance, we can now proceed to publication! I will do some final checks on your paper, but in the meantime please archive the software repo (e.g., at Zenodo) and provide the DOI here. And, thanks for your patience with this review.

Hi @kyleniemeyer I have archived the software repo - the DOI is 10.5281/zenodo.4555201
