### 2021-02-23 6ccd377
Another fix for _general_purpose_equality_check
When x or y is a string, the recursive check can get stuck in an
infinite loop because iterating a single character string yields the
same string again.
