### 2021-02-23 f14f089
Remove saving annotated images in IFT
Not sure if this is a useful feature.
