### 2021-02-23 382bb45
Add guards to prevent querying outside solution domain
