### 2021-02-23 2424bf3
Fix some oversights in young_laplace_guess
