+++ opendrop/fit/younglaplace/guess.py
@@ -1,15 +1,11 @@
-import math
 from typing import Tuple
 
 import numpy as np
 
+from opendrop.utility.misc import rotation_mat2d
+
 from ..pendant import find_pendant_apex
 from .types import YoungLaplaceParam
-
-
-# Math constants.
-PI = math.pi
-NAN = math.nan
 
 
 def young_laplace_guess(data: Tuple[np.ndarray, np.ndarray]) -> tuple:
@@ -17,10 +13,7 @@ def young_laplace_guess(data: Tuple[np.ndarray, np.ndarray]) -> tuple:
 
     apex, radius, rotation = find_pendant_apex(data)
 
-    r, z = np.array([[np.cos(rotation), -np.sin(rotation)],
-                     [np.sin(rotation),  np.cos(rotation)]]) \
-           @ data
-
+    r, z = rotation_mat2d(-rotation) @ (data - np.reshape(apex, (2, 1))) 
     bond = _bond_selected_plane(r, z, radius)
 
     params[YoungLaplaceParam.BOND] = bond
@@ -37,9 +30,9 @@ def _bond_selected_plane(r: np.ndarray, z: np.ndarray, radius: float) -> float:
     z_ix = np.argsort(z)
     if np.searchsorted(z, 2.0*radius, sorter=z_ix) < len(z):
         lower, upper = np.searchsorted(z, [1.95*radius, 2.05*radius], sorter=z_ix)
-        radii = np.abs(r[lower:upper+1])
-        r = radii.mean()/radius
-        bond = 0.1756 * r**2 + 0.5234 * r**3 - 0.2563 * r**4
+        radii = np.abs(r[z_ix][lower:upper+1])
+        x = radii.mean()/radius
+        bond = max(0.10, 0.1756 * x**2 + 0.5234 * x**3 - 0.2563 * x**4)
     else:
         bond = 0.15
 
