### 2021-02-23 bc119ca
Always save timeline.csv
