### 2021-02-23 #2604 @kyleniemeyer
Congrats @jdber1 on your article's publication in JOSS! And thank you for your patience here.

Many thanks to @rajeshrinet and @FelixKratz for reviewing this.

(It looks like the PDF has not yet built, but should show up momentarily.)
